### 2022-12-02 #4979 @lucydot
Hi @Kevin-Mattheus-Moerman this is outside of my expertise as I don't work with biological / molecular systems or imaging techniques. That said, if there is no-one else more suited I am happy to edit (I can edit one of the two papers you have invited me to).

As a note, when I was having a look at the software I found a related publication: https://www.cell.com/biophysj/fulltext/S0006-3495(17)34144-9 -  @MJWester is this an earlier iteration of the software presented here? If so, you will need to outline the changes that have been made since, so that we can assess whether the criteria (e.g. "scholarly effort" ) has been met since it was previously published. Please also, if you have not done so, include references to any other previous publications relating to the software.
