### 2022-12-05 #4979 @MJWester
Sorry, I was out-of-town for a few days.  The URL reference above to "A MATLAB-based Instrument Control Package for Fluorescence Imaging" is a different software that is dedicated to interfacing with instruments collecting super-resolution image data like microscopes, stages, lasers, etc., while SMITE is dedicated to dedicated to analyzing super-resolution image data that has been already collected, not necessarily using the instrument control software.

I will get back to you about reviewers.
