### 2022-12-07 2a4aa4d
Setup the possibility of BaGoL ROIs.
