### 2022-12-07 3d60578
Added ROIs as an argument to hierBaGoL_run.
