### 2022-12-20 3448a96
Add .fig plot to bivariateRipley_ROIcombined.
