### 2022-12-20 02d03c3
Make new figure visible!
