### 2022-12-27 f19b7af
pply BaGoL on a ROI basis if desired.
