+++ paper/paper.md
@@ -63,8 +63,11 @@ structures at the nanometer scale by fluorescently tagging the structures of
 interest, elucidating details of the biological behavior observed.
 
 Two common techniques are single-molecule localization microscropy (SMLM),
+[@Lidke_article:2005; @Betzig_article:2006; @Rust_article:2006;
+@Hell_article:2007; @vandeLinde_article:2011; @Fazel_article:2022]
 which is used to produce 2D or 3D super-resolution images of static or nearly
-static structures, and single-particle tracking (SPT), which follows the time
+static structures, and single-particle tracking (SPT), [@Shen_article:2017]
+which follows the time
 course of one or a very small number of moving tagged molecules.  SMLM often
 involves distributions of particles at medium to high density, while SPT works
 in a very low density domain.  These procedures all require intensive numerical
@@ -105,12 +108,13 @@ SMITE is a tool designed to be used by researchers and upper level students
 interested in fluorescence single molecule imaging and applications.  Parts of
 it have already been or are in the process of being published, e.g., frame
 connection [@Schodt_article:2021], drift correction [@Wester_article:2021],
-Bayesian grouping of localizations [@Fazel_article:2022].  Applications are
+Bayesian grouping of localizations [@Fazel_article:2022a].  Applications are
 described in [@Mazloom-Farsibaf_article:2021; @Bailey_article:2022].
 Typical raw image data can be found in [@Pallikkuth_data:2018].
 
-![SMITE GUIs for making movies from SPT trajectories, SMLM analysis, channel
-registration, and inspection of results contained in SMD
+![SMITE GUIs for (upper left) making movies from SPT trajectories, (upper
+right) SMLM analysis, (lower left) channel registration, and (lower right)
+inspection of results contained in SMD
 structures.\label{fig:smite_overview}](smite_overview.pdf){ width=100% }
 
 # Acknowledgements
