+++ paper/paper.bib
@@ -1,3 +1,120 @@
+# Early SMLM references.
+
+# Super-resolution microscopy references.
+@ARTICLE{Lidke_article:2005,
+   AUTHOR = {Keith A. {Lidke} and Bernd {Rieger} and Thomas M. {Jovin} and
+             Rainer {Heintzmann}},
+   TITLE = {{Superresolution by localization of quantum dots using blinking
+             statistics}},
+   JOURNAL = {Optics Express},
+   VOLUME = 13,
+   NUMBER = 18,
+   YEAR = 2005,
+   PAGES = {7052--7062},
+   DOI = {10.1364/OPEX.13.007052}
+   }
+
+# PALM (photoactivated localization microscopy).
+# Quantum dots or 50 nm luminescent gold beads used as fiducial markers.
+@ARTICLE{Betzig_article:2006,
+   AUTHOR = {Eric {Betzig} and George H. {Patterson} and Rachid {Sougrat} and
+             O. Wolf {Lindwasser} and Scott {Olenych} and Juan S. {Bonifacino}
+             and Michael W. {Davidson} and Jennifer {Lippincott-Schwartz} and
+             Harald F. {Hess}},
+   TITLE = {{Imaging Intracellular Fluorescent Proteins at Nanometer
+             Resolution}},
+   JOURNAL = {Science},
+   VOLUME = 313,
+   NUMBER = 5793,
+   MONTH = {September 15},
+   YEAR = 2006,
+   PAGES = {1642--1645},
+   DOI = {10.1126/science.1127344}
+   }
+
+# STORM (STochastic Optical Reconstruction Microscopy).
+# Fluorescent beads used as fiducial markers.
+@ARTICLE{Rust_article:2006,
+   AUTHOR = {Michael J {Rust} and Mark {Bates} and Xiaowei {Zhuang}},
+   TITLE = {{Sub-diffraction-limit imaging by stochastic optical reconstruction
+             microscopy (STORM)}},
+   JOURNAL = {Nature Methods},
+   VOLUME = 3,
+   NUMBER = 10,
+   MONTH = {August 9},
+   YEAR = 2006,
+   PAGES = {793--796},
+   PMID = {16896339},
+   PMCID = {PMC2700296},
+   DOI = {10.1038/nmeth929}
+   }
+
+@ARTICLE{Hell_article:2007,
+   AUTHOR = {Stefan W. {Hell}},
+   TITLE = {{Far-Field Optical Nanoscopy}},
+   JOURNAL = {Science},
+   VOLUME = 316,
+   NUMBER = 5828,
+   MONTH = {May 25},
+   YEAR = 2007,
+   PAGES = {1153--1158},
+   DOI = {10.1126/science.1137395},
+   }
+
+# dSTORM (direct STochastic Optical Reconstruction Microscopy)
+# Fluorescent beads or semiconductor nanocrystals used as fiducial markers.
+@ARTICLE{vandeLinde_article:2011,
+   AUTHOR = {Sebastian {van de Linde} and Anna {L\:oschberger} and Teresa
+             {Klein} and Meike {Heidbreder} and Steve {Wolter} and Mike
+             {Heilemann} and Markus {Sauer}},
+   TITLE = {{Direct stochastic optical reconstruction microscopy with standard
+             fluorescent probes}},
+   JOURNAL = {Nature Protocols},
+   VOLUME = 16,
+   NUMBER = {6(7)},
+   MONTH = {June},
+   YEAR = 2011,
+   PAGES = {991--1009},
+   PMID = {21720313},
+   DOI = {10.1038/nprot.2011.336}
+   }
+
+@ARTICLE{Fazel_article:2022,
[…]
