### 2023-01-04 3eea07b
Fixed a bug with combining label 2 BaGoL ROIs.
