### 2023-01-09 caed5ff
Fix a bug in the channel registration unit test
