### 2023-01-10 9263e90
Update the channel registration example
