### 2023-01-10 23d2917
Update the basic SPT example
