### 2023-01-10 af16735
Fix a problem with Verbose level.
