+++ MATLAB/examples/Example_LocalizeData.m
@@ -1,11 +1,11 @@
 %% Demonstrates LocalizeData Class
 
 %create a test dataset
-B=smi_sim.GaussBlobs.genRandomBlobImage();
-B=poissrnd(B);
+B = smi_sim.GaussBlobs.genRandomBlobImage();
+B = poissrnd(B);
 
 %setup SMF
-SMF=smi_core.SingleMoleculeFitting()
+SMF = smi_core.SingleMoleculeFitting()
 
 %setup LocalizeData object
 LD = smi_core.LocalizeData(B, SMF)
@@ -14,10 +14,5 @@ LD = smi_core.LocalizeData(B, SMF)
 [SMD] = LD.genLocalizations();
 
 %Set Verbose to give color overlay output
-LD.Verbose=2
+LD.Verbose = 3;
 [SMD] = LD.genLocalizations();
-
-    
-
-
-
