### 2023-01-11 8a5ddd2
Update README and paper.  Add SMITEclasses description.
