+++ SMITEclasses.txt
@@ -1,0 +1,43 @@
+SMITE classes
+-------------
+MATLAB/+smi:
+   @BaGoL/			Bayesian Grouping of Localizations
+   @Publish/			batch processing of SR data
+   @SMLM/			single molecule localization microscopy
+   @SPT/			single-particle tracking analysis
+MATLAB/+smi_cluster:
+   @Clustering/			clustering algorithms (DBSCAN, Voronoi, H-SET)
+   @PairCorrelation/		pair correlation (auto- and cross-) on ROIs
+   @StatisticsClustering/	clustering statistical analyses
+MATLAB/+smi_core:
+   @ChannelRegistration/	channel registration
+   @DataToPhotons/		convert raw data to photons
+   @DriftCorrection/		drift correction on 2D/3D data
+   @FRC/			Fourier Ring Correlation for image resolution
+   @FindROI/			find/collate subregions from a 2D image stack
+   @FrameConnection/		frame connection
+   @LoadData/			load raw microscope data from .mat/.h5 files
+   @LocalizeData/		find localizations in raw data
+   @SingleMoleculeData/		define SMD structure
+   @SingleMoleculeFitting/	define SMF structure
+   @Threshold/			threshold based on various SMLM properties
+   @TrackingResults/		define Tracking Results (TR) structure
+   GaussMLE.m			max likelihood estimate of 2D Gaussian blobs
+MATLAB/+smi_helpers:
+   @Filters/			filters useful for BaGoL operating on SMDs
+   @ROITools/			select ROIs from an image; save in a structure
+MATLAB/+smi_psf:
+   @PointSpreadFunction/	create and quantify point spread functions
+   @Zernike/			low-level Zernike polynomial functions
+MATLAB/+smi_sim:
+   @GaussBlobs/			generate 2D Gaussian blob images
+   @SimSMLM/			simulate SMLM data
+   @SimSPT/			simulate SPT data
+MATLAB/+smi_stat:
+   @ChangeDetection/		change detection analysis methods
+   @DiffusionEstimator/		diffusion estimation methods
+   @HMM/			hidden Markov model methods
+MATLAB/+smi_vis:
+   @GenerateImages/		general visualization of super-resolution data
+   @GenerateMovies/		generate movies
+   @InspectResults/		inspect super-resolution data
