### 2023-01-11 c769bd3
Save additional parameters in the BaGoL MAPN files.
