### 2023-01-11 2cfc2fa
Define OutputPixelSize/XStart/YStart properly for analysis.
