### 2023-01-13 fd0307e
Update batch tracking example with simulated data
