### 2023-01-13 98dfec8
Update HMM example to use simulated data
