+++ MATLAB/examples/Example_HMM.m
@@ -1,31 +1,104 @@
-% This script demonstrates the basic usage of smi_stat.HMM.
+% This script demonstrates the basic usage of smi_stat.HMM.  We start by
+% simulating some raw tracking data, reloading the raw data and tracking,
+% and finally running the tracking results through the HMM analysis
+% pipeline to search for dimers.  Simulated data and results will be saved
+% in freshly generated directories under smite/MATLAB/examples/spt/
 
-%% Reload results from smi.SPT tracking and search for dimer candidates.
+%% Simulate and save some SPT data in the format expected for real data.
+% Simulate some diffusing blobs and save simulated raw data.
+rng(12)
+DataParams.NDatasets = 10;
+SimParams.FrameSize = [64, 64];
+SimParams.NFrames = 1000;
+SimParams.ParticleDensity = 0.01; % particles / px^2, make 2X target since we split into 2 channels
+SimParams.D = 0.1; % px^2 / frame
+SimParams.KOffToOn = 0.95;
+SimParams.KOnToOff = 0.05;
+SimParams.KOnToBleach = 1e-3;
+SimParams.Intensity = 1000;
+SimParams.KDisconnect = 0.05;
+SimParams.InteractionDistance = 0.5; % pixels
+SimParams.InteractionProb = 0.5;
+SMITEPath = fileparts(which('setupSMITE.m'));
+DataDir = fullfile(SMITEPath, ...
+    'examples', 'example_data', 'spt', smi_helpers.genTimeString());
+if ~isfolder(DataDir)
+    mkdir(DataDir)
+end
+[~, SimParams, DataParams] = smi_sim.SimSPT.makeExampleSim(...
+    SimParams, DataParams, DataDir); % SimParams and DataParams padded w/ defaults
+
+%% Prepare an SMF structure for each channel.
+% NOTE: Channel 1 will not be transformed (it is the reference channel), so
+%       we should ensure SMFChannel1.Data.RegistrationFilePath is empty!
+PixelSize = 0.1; % must be set to true data values
+FrameRate = 20;
+SMFChannel1 = smi_core.SingleMoleculeFitting;
+SMFChannel1.Data.AnalysisID = 'Channel1';
+SMFChannel1.Data.FileDir = DataDir;
+SMFChannel1.Data.DataROI = [1, 1, ...
+    SimParams.FrameSize(1), SimParams.FrameSize(2)]; % left half of data [YStart, XStart, YEnd, XEnd]
+SMFChannel1.Data.RegistrationFilePath = '';
+SMFChannel1.Data.PixelSize = PixelSize;
+SMFChannel1.Data.FrameRate = FrameRate;
+SMFChannel1.Fitting.PSFSigma = SimParams.PSFSigma;
+SMFChannel1.Tracking.MaxFrameGap = 20;
+SMFChannel2 = copy(SMFChannel1);
+SMFChannel2.Data.AnalysisID = 'Channel2';
+SMFChannel2.Data.DataROI = [1, 1 + SimParams.FrameSize(2), ...
+    SimParams.FrameSize(1), 2*SimParams.FrameSize(2)]; % right half of data
+SMFChannel2.Data.RegistrationFilePath = '';
+
+%% Prepare an SPT class object for each channel and then track.
+SPTChannel1 = smi.SPT(SMFChannel1, false);
+SPTChannel1.GenerateMovies = false;
+SPTChannel1.GeneratePlots = false;
+SPTChannel2 = smi.SPT(SMFChannel2, false);
+SPTChannel2.GenerateMovies = false;
+SPTChannel2.GeneratePlots = false;
+
+%% Loop through our data files and track one at a time.
+FileList = dir(fullfile(DataDir, 'Data_*.mat'));
[…]
