+++ MATLAB/examples/Example_SPTBatch.m
@@ -9,7 +9,7 @@ DataParams.NDatasets = 10;
 SimParams.FrameSize = [64, 64];
 SimParams.NFrames = 1000;
 SimParams.ParticleDensity = 0.01; % particles / px^2, make 2X target since we split into 2 channels
-SimParams.D = 0.1; % px^2 / s
+SimParams.D = 0.1; % px^2 / frame
 SimParams.KOffToOn = 0.95;
 SimParams.KOnToOff = 0.05;
 SimParams.KOnToBleach = 1e-3;
