### 2023-01-23 94ff1c4
Various updates on the doc, paper and examples.
