+++ paper/paper.bib
@@ -311,3 +311,52 @@ H. Shen, L. J. Tauzin, R. Baiyasi, W. Wang, N. Moringo, B. Shuang, and C. F. Lan
    URL = {https://digitalrepository.unm.edu/physics_data/3/},
    DOI = {10.25827/CS2A-DH13}
 }
+
+# =============================================================================
+
+# Instrument control
+@ARTICLE{Pallikkuth_article:2018,
+   AUTHOR = {Sandeep {Pallikkuth} and Marjolein {Meddens} and Mohamad {Fazel}
+             and Hanieh {Farsibaf} and Farzin {Farzam} and Michael {Wester} and
+             Keith {Lidke}},
+   TITLE = {{A MATLAB-based Instrument Control Package for Fluorescence
+             Imaging}},
+   JOURNAL = {Biophysical Journal},
+   VOLUME = 114,
+   NUMBER = {3, Supplement 1},
+   MONTH = {February 2},
+   YEAR = 2018,
+   PAGES = {532a},
+   DOI = {10.1016/j.bpj.2017.11.2912"}
+   }
+
+# =============================================================================
+
+# Other integrated software
+@ARTICLE{Ries_article:2020,
+   AUTHOR = {Jonas {Ries}},
+   TITLE = {{SMAP: a modular super-resolution microscopy analysis platform for
+             SMLM data}},
+   JOURNAL = {Nature Methods},
+   VOLUME = 17,
+   MONTH = {September},
+   YEAR = 2020,
+   PAGES = {370--372},
+   URL = {https://www.nature.com/articles/s41592-020-0938-1},
+   DOI = {10.1038/s41592-020-0938-1}
+}
+
+@ARTICLE{Edelstein_article:2014,
+   AUTHOR = {Arthur D {Edelstein} and Mark A {Tsuchida} and Nenad {Amodaj} and
+             Henry {Pinkard} and Ronald D {Vale} and Nico {Stuurman}},
+   TITLE = {{Advanced methods of microscope control using $\mu$Manager
+             software}},
+   JOURNAL = {Journal of Biological Methods},
+   VOLUME = 1,
+   NUMBER = 2,
+   MONTH = {November},
+   YEAR = 2014,
+   PAGES = e11, 
+   URL = {https://polscientific.com/jbm/index.php/jbm/article/view/36},
+   DOI = {10.14440/jbm.2014.36}
+}
