### 2023-01-27 2697b5f
Update paper and a few unitTests.
