+++ paper/paper.bib
@@ -267,6 +267,22 @@ H. Shen, L. J. Tauzin, R. Baiyasi, W. Wang, N. Moringo, B. Shuang, and C. F. Lan
 # =============================================================================
 
 # Used SMITE tools.
+@ARTICLE{FrancoNitta_article:2021,
+   AUTHOR = {Carolina {Franco Nitta} and Ellen W {Green} and Elton D {Jhamba}
+             and Justine M {Keth} and Ira\'is {Ortiz-Caraveo} and
+             Rachel M {Grattan} and David J {Schodt} and Aubrey C {Gibson} and
+             Ashwani {Rajput} and Keith A {Lidke} and Bridget S {Wilson} and
+             Mara P {Steinkamp} and Diane S {Lidke}},
+   TITLE = {{EGFR transactivates RON to drive oncogenic crosstalk}},
+   JOURNAL = {eLife},
+   VOLUME = 10,
+   YEAR = 2021,
+   PAGES = {e63678: 1--27},
+   URL = {https://doi.org/10.7554/eLife.63678},
+   DOI = {10.7554/eLife.63678}
+}
+
+# Used SMITE tools.
 @ARTICLE{Mazloom-Farsibaf_article:2021,
    AUTHOR = {Hanieh {Mazloom Farsibaf} and Farzin {Farzam} and Mohamadreza
              {Fazel} and Michael J. {Wester} and Marjolein B. M. {Meddens} and
