### 2023-01-31 d5ec708
Fix problems with some unitTests; update documentation.
