+++ MATLAB/+smi_vis/@GenerateImages/circleImage_unitTest.m
@@ -49,8 +49,8 @@ ColorMapMagenta = repmat([1, 0, 1], NEmittersPerChannel, 1);
 ColorMap = [ColorMapGreen; ColorMapMagenta];
 [CircleImage, CircleImageRGB, SRImageZoom] = ...
     smi_vis.GenerateImages.circleImage(SMDAll, ColorMap, [], 16);
-imshow(CircleImage);
-imshow(CircleImageRGB);
+figure; imshow(CircleImage);
+figurecdimshow(CircleImageRGB);
 Success = 1;
 
 
