### 2023-02-10 6c1f516
Fix a bug in smi_core.LocalizeData unit test
