+++ MATLAB/+smi_core/@LocalizeData/unitTest.m
@@ -39,7 +39,6 @@ SMD.XSize = FrameSizeFull;
 
 % Generate an SMF structure.
 SMF = smi_core.SingleMoleculeFitting;
-SMF.Data.CameraType = 'SCMOS';
 SMF.BoxFinding.BoxSize = 10;
 SMF.Fitting.PSFSigma = SMD.PSFSigma;
 
@@ -51,12 +50,9 @@ LD = smi_core.LocalizeData(ScaledData, SMF, 3);
 % other (this isn't meant to be an exact check of individual fields).
 NEmitters = 5;
 Success(1) = ...
-    (all(SMD.FrameNum==repelem((1:NFrames).', NEmitters)) ...
-    && (numel(SMD.X)==NEmitters*NFrames) ...
+    (all(SMDout.FrameNum==repelem((1:NFrames).', NEmitters)) ...
+    && (numel(SMDout.X)==NEmitters*NFrames) ...
     && (numel(SMDout.X)==sum(SMDPreThresh.ThreshFlag==0)));
-%Success(1) = ...
-%    (all(SMDPreThresh.FrameNum==repelem((1:NFrames).', NEmitters)) ...
-%    && (numel(SMDPreThresh.X)==NEmitters*NFrames) ...
-%    && (numel(SMDout.X)==sum(SMDPreThresh.ThreshFlag==0)));
+
 
 end
