+++ MATLAB/+smi_cluster/@PairAnalysis/doAnalysis.m
@@ -1,0 +1,178 @@
+function [results_pcc, resultsRC_pcc, results_c, results_cs, results_ls, ...
+          results_o1, results_o2] =                                      ...
+   doAnalysis(n_ROIs, RoI, ROI_sizes, desc, particles, results_dir,      ...
+              options, PixelSize, HistBinSize, RmaxAxisLimit, E, minPts, ...
+              PlotNonOverlap, Color)
+% Dispatches analyses to various helper functions depending options provided.
+%
+% INPUTS:
+%    n_ROIs        number of ROIs
+%    RoI           data structure containing x and y coordinates per ROI
+%    ROI_sizes     ROI sizes (nm)
+%    desc          string identifying the analysis
+%    particles     string array describing the two particles
+%    results_dir   output directory
+%    options       string array of strings specifying analyses to be performed:
+%                      'combined'    the analysis is combined over all ROIs
+%                      'plotting'    plots are to be produced.
+%                      'BiStats'     pairwise mutual distances and bivariate
+%                                    Ripley's statistics for each ROI
+%                      'Clustering'  clusters for each label per ROI
+%                      'Clustering2' C2C nearest neighbor distances between
+%                                    label 1/label 2 clusters per ROI and
+%                                    combined over all ROIs
+%                      'LocSep2'     NN localization distances between L1 & L2
+%                      'Overlap1'    overlap between L1 clusters/L2 locs.
+%                      'Overlap2'    overlap between L2 clusters/L1 locs.
+%                      'PairCorr'    pair correlation per ROI and combined over
+%                                    all ROIs
+%                      'Plot2'       2D plot per ROI
+%    PixelSize     conversion factor conversion factor (nm)
+%    HistBinSize   number of pixels per bin to collect correlation statistics;
+%                  a good guess is the PixelSize
+%    RmaxAxisLimit sets r axis limit for pair correlation plots if > 0 (nm)
+%    E             2-element array defining epsilon for clustering each label
+%    minPts        2-element array defining min cluster size for each label
+%    PlotNonOverlap if true, plot non-overlapping locs as black dots, otherwise
+%                  omit them from the plot
+%    Color         label 1 and label 2 colors on display
+%
+% OUTPUTS:
+%    desc_results.mat containing the various results_ cell arrays for one cell
+%    Various results containers from the called helper functions in case the
+%    user wants to have more details:
+%                       bivariate statistics
+%          Also, figures *_ROI*_L1/2_pairwiseCDF/PDF compared to a random dist.
+%                *_ROI*_L1,L2_pairwisePDF2/CDF2      2-label PDF/CDF
+%                *_ROI*_bivripley                    bivariate Ripley
+%       results_pcc     pair cross-correlation
+%                          results_pcc{1:n_ROIs}
+%                             .G       2D pair-correlation function values
+%                             .r       radius values (nm)
+%                             .g       angularly averaged pair corr. function
+%                             .dg      errors on angularly averaged g
+%                             .model   model calculated at estimated value
+%                             ....     various model results
+%          Also, figures *_ROI*_crosscorr (ROIwise pairwise cross-correlations)
+%       resultsRC_pcc   ROIs combined pair cross-correlation
+%                          results_pcc{1:n_ROIs}
+%                             see results_pcc (ROI combined pairwise_crosscorr)
+%          Also, figures *_RC_crosscorrR
[…]

[…]
