### 2023-03-03 dfc25c0
Add some more examples based on scripts developed over time.
