+++ MATLAB/examples/simplePairCorr.m
@@ -1,0 +1,287 @@
+% Work flow: (note results are put in the subdirectory 'Analysis')
+% NOTE: smite should be on your MATLAB path.
+%
+%   Set important parameters:
+%      ROI_sizes, Pixel2nm, start_datadir
+%   Define the ROIs:
+%            _ResultsStruct.mat (or _Results.mat) -> _ROIs.mat
+%      (each image [ResultsStruct or Results file] produces a separate ROIs
+%       file containing all the ROIs for that image; the idea is to do all the
+%       ROI selection early on so it need not be repeated---each image can have
+%       multiple ROIs; note that this deals with files in a single directory)
+%   Analyze the ROIs one-by-one:
+%            _ROIs.mat -> _results.mat, etc.
+%      (compute results for each ROI defined by the selected files located in
+%       the same directory)
+%   Analyze a group of files of ROIs all together:
+%            multiple _ROIs.mat
+%                 -> ALL_ and analysis_ALL_*.mat
+%      (compute combined results for the ROIs defined by the selected files
+%       located in the same directory)
+%
+% Avoid spaces in filenames and conditions!!!
+%
+% See doAnalysis    for analyses to perform:
+%     doBiStats     for mutual pairwise distance and bivariate Ripley's plots
+%     doClustering  for cluster parameters: E and minPts
+%     doClusterSep2 for cluster separation plots
+%     doLocSep2     for nearest neighbor localization separation plots
+%     doOverlap1    for overlaps between label 1 clusters and label 2 locs
+%     doOverlap2    for overlaps between label 2 clusters and label 1 locs
+%     doPairCorr    for pair correlation settings
+%     doPlot2       for 2-color plots
+%
+% Output figure naming conventions:
+%    *_ROI*_L1/2_algorithm            clustering by algorithm
+%    *_ROI*_L1/2_pairwiseCDF/PDF      NN distances compared to a random dist.
+%    *_ROI*_L1,L2_pairwisePDF2/CDF2   NN distances 2-label PDF/CDF
+%    *_ROI*_bivripley                 bivariate Ripley's
+%    *_ROI*_crosscorr                 ROIwise pairwise cross-correlations
+%    *_RC_crosscorrR                  ROI combined pairwise cross-correlation
+%    *_cL1_lL2                        L1 clusters, L2 localizations overlap
+%    *_cL2_lL1                        L2 clusters, L1 localizations overlap
+%    *_ROI*_L1+L2                     2-label plots of localizations
+%    *_cs_RC_PDF/CDF                  ROI combined 2-label cluster separations
+%    *_ls_ROI*                        histogram per ROI of loc. L1+L2 seps.
+%    *_ls_RC_PDF/CDF                  ROI combined localization L1+L2 seps.
+
+%% --- Set important parameters
+% Collect the set of files expressing each label together.
+
+PA = smi_cluster.PairAnalysis();
+
+%start_datadir = '/mnt/nas/cellpath/Genmab/Data/20221119_CHO_HAEGFR_antiHA_antiC1Q/PairwiseCorr';
+start_datadir = '.';
+Analysis = 'Analysis';
+
+particles = {'L1', 'L2'};     % label short descriptors
+Color = ['g', 'm'];           % label 1 and 2 colors on display
+%Pixel2nm = 108.018;           % pixels to nm [TIRF]
[…]
