+++ MATLAB/examples/simpleROIcluster.m
@@ -1,0 +1,294 @@
+% Perform cluster analysis for comparison of experimental conditions.
+% NOTE: sma-cluster and PlotSpread should be on your MATLAB path.
+%
+% Work flow: (note results are put in the subdirectory 'Analysis')
+%   Set important parameters:
+%            ROI_sizes, Pixel2nm, start_datadir, oneROI
+%      ALWAYS run this first MATLAB section before running the later sections
+%      so that necessary parameters are set.
+%   Define the ROIs:
+%            _ResultsStruct.mat (or _Results.mat) -> _ROIs.mat
+%      (each image [ResultsStruct or Results file] produces a separate ROIs
+%       file containing all the ROIs for that image; the idea is to do all the
+%       ROI selection early on so it need not be repeated---each image can have
+%       multiple ROIs; note that this deals with files in a single directory.
+%       Many time lab members select the ROIs from the SR _Results.mat files,
+%       To transfer these to BaGoL ROIs (necessary, as the ROI files are
+%       self-contained, meaning they contain the actual coordinates of the
+%       localizations in each ROI), it is necessary to transfer the ROI regions
+%       to BaGoL localizations.  See below.)
+%   POSSIBLY, define the BaGoL ROIs from the previous ROIs and BaGoL results
+%            _ROIs.mat + _BaGoL_Results.mat -> _BaGoL_ROIs.mat
+%            _ROIs.mat + MAPN_*.mat         -> _BaGoL_ROIs.mat
+%      (the two sets of files should be in corresponding order; the BaGoL
+%       coordinates replace the coordinates in the original ROI files)
+%   POSSIBLY, combine individually processed BaGoL ROIs into a single _ROIs.mat
+%      file using the _ROIs.mat file that was used to define the ROIs
+%      originally from the SR data
+%            _ROIs.mat + _Results_ROI_*_.mat -> _Results.mat
+%   POSSIBLY, filter out some ROIs.
+%            _ROIs.mat -> _filtered_ROIs.mat
+%   Statistics for a single condition:
+%            _ROIs.mat -> ALL_*
+%      (compute statistics for a single condition consisting of a number of
+%       images all located in the same directory.  Need to define
+%       algorithm_range, minPts_range (N) and E_range (epsilon).)
+%   Combine results from one or more conditions for further processing:
+%            multiple ALL_*_results.mat -> [condition]_results.mat
+%      (multiple instantiations of a single condition located in multiple
+%       directories are collected into a combined instance in a specified
+%       directory.)
+%   Combined statistics for two or more conditions:
+%            multiple [condition]_results_mat -> 
+%      (multiple conditions are plotted on the same graph and placed in a
+%       subdirectory for this particular analysis.  The filenames are assumed
+%       to have the structure:
+%          experimentalConditions#analysisConditions_results.mat
+%       where experimentalConditions contains no #.  Statistics for a single
+%       condition produces filenames in this format, so can be copied directly.
+%       A typical example of a renamed ALL_*_results file is:
+%          RGY#DBSCAN_N=3,E=50_results.mat
[…]
