### 2023-03-08 6f9e83d
saveBaGoL overlay image colors indicated in filename now.
