### 2023-03-09 #4979 @jgostick
It's looking much better, though still lacking two rather large pieces: tests and documentation.

## Tests

The reviewers are asked to look for that, and I think it's pretty much mandatory for a good software package to have some tests.  Mathworks [has a repo](https://github.com/mathworks/ci-configuration-examples#github-actions) explaining how to setup automatated tests on pushes to a repo.  @Kevin-Mattheus-Moerman, I'm not sure how strongly we enforce this testing requirement?  I think at least the pack should include a file called "run_some_tests.m" which is a script that calls functions with known arguments to make sure the right values are returned.

BTW, @MJWester, I learned to add units tests to my packages a while back and it's been a life saver.  I have caught **many** bugs and problems that I created by accident when changing something in our code base.  So knowing that any changes are probably ok really relieves anxiety when making changes, especially bigger ones.  My point it that writing tests is worth the effort.

## Docs
I still cant see any indication of **how to use** this package?  Shouldn't there be a website with some examples?  Given that it's a microscopy package, it would be easy to make it look nice with lots of images of before and after applying some function, right?  As as former Matlab user myself, I know that I really appreciated their detailed help files, so I am guessing that your potential users would also expect something equivalent for this toolbox?
