### 2023-03-11 #4979 @Kevin-Mattheus-Moerman
@MJWester please work on the following based on @jgostick's comments

- [ ] Add automated testing, e.g. a script running several test codes. Note, I am the author of the MATLAB package [GIBBON](https://github.com/gibbonCode/GIBBON/), and it features something similar, a script called [testGibbon.m](https://github.com/gibbonCode/GIBBON/blob/master/lib/testGibbon.m), it represents perhaps quite a complicated version of what is suggested, but perhaps you can use it for inspiration. It simply looks in a folder containing the text scripts and runs them all (or a subset as required).

- [ ] Extend the documentation to cover all core functionality. Once again you can check GIBBON, which once again may be more complex than what you'd like to implement, but it features a [`docs` folder]() containing formatted demos and help files. For instance this file:
https://github.com/gibbonCode/GIBBON/blob/master/docs/HELP_im2patch.m
Is the source for this documentation HTML (after running `gpublish HELP_im2patch.m`):
https://github.com/gibbonCode/GIBBON/blob/master/docs/html/HELP_im2patch.html
Which I render on the website here:
https://www.gibboncode.org/html/HELP_im2patch.html
Note MATLAB's _publish_ also allows you to create pdf files if that is preferred over HTML. Either way, please choose an approach to include and render functionality documentation. If the GIBBON approach is of interest I'd be happy to help answer questions.

- [ ] As requested please include information on **how to use** this package. You can consider, like with GIBBON, to render a github pages website showing the documentation. Also such a site, as well as the README, should have clear installation, testing, and use instructions, and I recommend adding a "Getting started" section too.

I'll await your comments. If you are happy to implement the above we can pause the submission again until you are finished. If not we'll need to retract this submission due to lack of testing/documentation/use instructions.
