+++ doc/DetailedOverview.md
@@ -47,9 +47,9 @@ Corresponding examples are presented in:
 
 In all the examples and throughout ***smite***, three main data
 structures are used to organize the analysis parameters and data:
-- [SMF](CoreFunctionality/SMF) (Single Molecule Fitting), which
+- [SMF](DataStructures/SMF) (Single Molecule Fitting), which
   uniquely and completely defines the data analysis parameters;
-- [SMD](CoreFunctionality/SMD) (Single Molecule Data), which stores
+- [SMD](DataStructures/SMD) (Single Molecule Data), which stores
   the data results;
-- [TR](CoreFunctionality/TR) (Tracking Results), which is a variation
+- [TR](DataStructures/TR) (Tracking Results), which is a variation
   of SMD used for tracking data results.
