+++ MATLAB/+smi_sim/README.md
@@ -1,0 +1,4 @@
++smi_sim is the namespace for the simulation classes of ***smite***:
+- (@GaussBlobs)[@GaussBlobs/README.md]: generate 2D Gaussian blob images
+- (@SimSMLM)[@SimSMLM/README.md]:       simulate SMLM data
+- (@SimSPT)[@SimSPT/README.md]:         simulate SPT data
