+++ MATLAB/+smi_vis/README.md
@@ -1,0 +1,13 @@
++smi_vis is the namespace for the visualization functions and classes of
+***smite***:
+- (@GenerateImages)[@GenerateImages/README.md]:
+  general visualization of super-resolution data
+- (@GenerateMovies)[@GenerateMovies/README.md]:
+  generate movies
+- (@InspectResults)[@InspectResults/README.md]:
+  inspect super-resolution data
+
+Functions:
+- contrastStretch: scales images to the range MinMax
+- plotCDFs:        plots multiple CDFs on the same plot.
+- plotImageStack:  plots a stack of images contained in a 3D array
