+++ doc/FileFormats/CalibrationFile.md
@@ -1,3 +1,5 @@
+### CalibrationFile
+
 Example camera calibration file:
 
 ```
