+++ MATLAB/+smi/README.md
@@ -1,5 +1,7 @@
+### +smi
+
 +smi is the namespace for the core user top-level classes of ***smite***:
-- (@BaGoL)[@BaGoL/README.md]:     Bayesian Grouping of Localizations
-- (@Publish)[@Publish/README.md]: batch processing of SR data
-- (@SMLM)[@SMLM/README.md]:       Single Molecule Localization Microscopy
-- (@SPT)[@SPT/README.md]:         Single-Particle Tracking analysis
+- [@BaGoL](@BaGoL/README.md):     Bayesian Grouping of Localizations
+- [@Publish](@Publish/README.md): batch processing of SR data
+- [@SMLM](@SMLM/README.md):       Single Molecule Localization Microscopy
+- [@SPT](@SPT/README.md):         Single-Particle Tracking analysis
