+++ doc/FileFormats/H5.md
@@ -1,3 +1,5 @@
+### .h5 file
+
 Example H5 file structure:
 
 ```
