+++ MATLAB/+smi_psf/README.md
@@ -1,5 +1,7 @@
+### +smi_psf
+
 +smi_psf is the namespace for the Point Spread Function classes of ***smite***:
-- (@PointSpreadFunction)[@PointSpreadFunction/README.md]:
+- [@PointSpreadFunction](@PointSpreadFunction/README.md):
   create and quantify point spread functions
-- (@Zernike)[@Zernike/README.md]:
+- [@Zernike](@Zernike/README.md):
   low-level Zernike polynomial functions
