+++ MATLAB/+smi_core/README.md
@@ -1,27 +1,29 @@
-+smi is the namespace for the core lower-level classes of ***smite***:
-- (@ChannelRegistration)[@ChannelRegistration/README.md]:
+### +smi_core
+
++smi_core is the namespace for the core lower-level classes of ***smite***:
+- [@ChannelRegistration](@ChannelRegistration/README.md):
   channel registration
-- (@DataToPhotons)[@DataToPhotons/README.md]:
+- [@DataToPhotons](@DataToPhotons/README.md):
   convert raw data to photons
-- (@DriftCorrection)[@DriftCorrection/README.md]:
+- [@DriftCorrection](@DriftCorrection/README.md):
   drift correction on 2D/3D data
-- (@FRC)[@FRC/README.md]:
+- [@FRC](@FRC/README.md):
   Fourier Ring Correlation for image resolution
-- (@FindROI)[@FindROI/README.md]:
+- [@FindROI](@FindROI/README.md):
   find/collate subregions from a 2D image stack
-- (@FrameConnection)[@FrameConnection/README.md]:
+- [@FrameConnection](@FrameConnection/README.md):
   frame connection
-- (@LoadData)[@LoadData/README.md]:
+- [@LoadData](@LoadData/README.md):
   load raw microscope data from .mat/.h5 files
-- (@LocalizeData)[@LocalizeData/README.md]:
+- [@LocalizeData](@LocalizeData/README.md):
   find localizations in raw data
-- (@SingleMoleculeData)[@SingleMoleculeData/README.md]:
+- [@SingleMoleculeData](@SingleMoleculeData/README.md):
   define SMD structure
-- (@SingleMoleculeFitting)[@SingleMoleculeFitting/README.md]:
+- [@SingleMoleculeFitting](@SingleMoleculeFitting/README.md):
   define SMF structure
-- (@Threshold)[@Threshold/README.md]:
+- [@Threshold](@Threshold/README.md):
   threshold based on various SMLM properties
-- (@TrackingResults)[@TrackingResults/README.md]:
+- [@TrackingResults](@TrackingResults/README.md):
   define Tracking Results (TR) structure
-- (GaussMLE)[README_GaussMLE.md]:
+- [GaussMLE](README_GaussMLE.md):
   max likelihood estimate of 2D Gaussian blobs
