+++ MATLAB/+smi_core/@LoadData/README.md
@@ -1,306 +1,30 @@
-classdef LoadData < handle
-    % LoadData has methods to load raw microsope dataset from .mat or
-    % .h5 format files.
-    %
-    % INPUTS:
-    %   SMF - Single Molecule Fitting structure. The name and directory
-    %   location of the data file is obtained from fields
-    %   (SMF.Data.FileName and SMF.Data.FileDir) of SMF. If the fields are
-    %   non-existent or are empty, data file selection will be required
-    %   with a pop-up window.
-    %   varargin - input arguments helping to locate the 'data' to be
-    %   loaded from the data file. This varies for the different file
-    %   extensions (.mat and .h5). For details please see
-    %   documentation for the methods.
-    %
-    % OUTPUTS:
-    %   Data - data loaded from FileName, converted to type single
-    %   SMF - SMF structure modified by adding fields FileName and FileDir,
-    %   if not present.
-    %
-    % METHODS:
-    %   loadDataMat, loadDataH5
-    %
-    % CITATION:
-    %   Sandeep Pallikkuth, Lidke Lab, 2020
-    
-    properties (SetAccess = 'protected')
-        FileType           % Filetype determined from SMF.Data.FileName
-        FullFileName       % Full file path
-    end
-    
-    methods
-        function obj = LoadData()
-        % Constructor.
-        end
+### +smi_core/@LoadData
 
-        function [obj,Data,SMF] = loadRawData(obj,SMF,varargin)
-            % INPUT
-            %    SMF - single molecule fitting structure
-            %    varargin - input parameters, different for each file type:
-            %       mat: DatasetNum, number indicating the file in
-            %               SMF.Data.FileName cell array
-            %            
-            %       h5: DatasetIdx, index of dataset to be loaded from h5 file
-            %           ChannelIdx (optional), index of channel to be loaded, default is 1
-            % OUTPUT
-            %    Data - data loaded from FileName, converted to type single
-            %    SMF - single molecule fitting structure with FileName,
-            %    FileDir and ResultsDir added.
-            
-            if isfield(SMF.Data,'FileName')&&~isempty(SMF.Data.FileName)
-                % determining FileTye from SMF entries of FileName
-                if iscell(SMF.Data.FileName)
-                    obj.FileType=SMF.Data.FileName{1}(end-2:end);
-                else
-                    obj.FileType = SMF.Data.FileName(end-2:end);
-                end
-                obj.FullFileName=fullfile(SMF.Data.FileDir,SMF.Data.FileName);
-            else
-                % If Filename entry not present in SMF, option to choose
-                % the file/s
-                [filename, pathname]=uigetfile('Y:\*.mat;*.mat;*.h5','MultiSelect','on');
-                SMF.Data.FileName=filename;
-                SMF.Data.FileDir=pathname;
-                SMF.Data.ResultsDir=fullfile(pathname,'Results');
-                obj.FullFileName=fullfile(SMF.Data.FileDir,SMF.Data.FileName);
-                if iscell(SMF.Data.FileName)
-                    obj.FileType=SMF.Data.FileName{1}(end-2:end);
-                else
-                    obj.FileType = SMF.Data.FileName(end-2:end);
-                end
-            end
-            
-            switch obj.FileType
-                case 'mat'
-                    % loading images from .mat file/s
-                    % MatVarName, name of matlab variable containing the data
-                    MatVarName = SMF.Data.DataVariable;
-                    varargin = {varargin{:}, MatVarName};
-                    Data=smi_core.LoadData.loadDataMat(obj.FullFileName,varargin);
-                case '.h5'
-                    % loading images from .h5 file
-                    Data=smi_core.LoadData.loadDataH5(obj.FullFileName,varargin);
[…]
