+++ doc/Tutorials/BaGoL_EGFR_dSTORM.m
@@ -1,0 +1,166 @@
+%% Bayesian Grouping of Localizations (BaGoL) Example for dSTORM EGFR
+%  BaGoL is run for a region of the data using hierarchical  
+%  Bayes to find the distribution of the number of localizations per  
+%  emitter (Xi) from the data itself. In the second run of BaGoL, the entire 
+%  region is processed using the found distribution as an input.
+%
+% Requirements and Setup:
+%   1. Windows 64 bit OS
+%   2. MATLAB 2017b or higher versions
+%   3. Statistics and Machine Learning Toolbox
+%   4. BaGoL class
+%   5. Set MATLAB directory to BaGoL directory.
+%
+% Description of how to run...
+%   1. Set the parameters in the following section
+%   2. Set a cutoff for filtering too bright localizations 
+%   3. Select a region of data to learn Xi.
+%   4. Set the parameter 'DataROI' for the entire data set to be processed.
+%
+% Results include:
+%   Saved Results:
+%     SR_Im.png:                 Traditional super-resolution image. 
+%     Post-Im.png:               Posterior image or histogram image of the chain
+%                                (weighted average over all models).
+%     MAPN-Im.png:               MAPN image which is the image of localizations from the
+%                                most likely model. 
+%     Overlay_SR_Map.png:        Overlay of grayscale SR-image and color MAPN image.
+%     Overlay_SR_Post.png:       Overlay of grayscale SR-image and color posterior image. 
+%     Overlay_SR_Map_circle.png: Overlay of the SR & MAPN coordinates where 
+%                                every coordinate is represented by a circle  
+% 		                         located at the given location and a radius 
+%                                of double of the given precision.
+%     Xi.png:                    Number of localizations per emitter dist.
+%     NND.png:                   Histogram of nearest neighbor distances from
+%                                MAPN-coordinates. 
+%     BaGoL_X-SE.png:            Histogram of X-localization precisions after grouping. 
+%     BaGoL_Y-SE.png:            Histogram of Y-Localization precisions after grouping.
+%     LocsScatter-MAPN.fig:      Plot of time color-coded localizations and
+%                                MAPN-coordinates.
+%     MAPN.mat:                  Structure containing the MAPN-coordinates of emitters.
+%   Output available on work space:
+%     MAPN: Clusters information are stored in this property:
+%     MAPN.X: X-Centers (nm)
+%     MAPN.Y: Y-Centers (nm)
+%     MAPN.X_SE: X-Centers precisions (nm)
+%     MAPN.Y_SE: Y-Centers precisions (nm)
+%     MAPN.AlphaX: X-Drifts of clusters (nm/frame)
+%     MAPN.AlphaY: Y-Drifts of clusters (nm/frame)
+%     MAPN.AlphaX_SE: X-Drift precisions (nm/frame)
+%     MAPN.AlphaY_SE: Y-Drift precisions (nm/frame)
[…]
