### 2023-04-14 b107042
Some extended examples added + doc fixes.
