+++ MATLAB/examples/README.md
@@ -45,13 +45,13 @@ SMITE unit tests:
 - [smi_core.LocalizeData.unitTest](../+smi_core/@LocalizeData/unitTest.m)
 - [smi_core.Threshold.unitTest](../+smi_core/@Threshold/unitTest.m)
 - [smi_psf.PointSpreadFunction.unitTest](../+smi_psf/@PointSpreadFunction/unitTest) (does the following unit tests)
-> [smi_psf.PointSpreadFunction.crlbPSFPupil_unitTest](../+smi_psf/@PointSpreadFunction/crlbPSFPupil_unitTest.m)
-> [smi_psf.PointSpreadFunction.optimPSFZernike_unitTest](../+smi_psf/@PointSpreadFunction/optimPSFZernike_unitTest.m)
-> [smi_psf.PointSpreadFunction.oversamplePSFPupil_unitTest](../+smi_psf/@PointSpreadFunction/oversamplePSFPupil_unitTest.m)
-> [smi_psf.PointSpreadFunction.phaseRetrieval_unitTest](../+smi_psf/@PointSpreadFunction/phaseRetrieval_unitTest.m)
-> [smi_psf.PointSpreadFunction.psfROIStack_unitTest](../+smi_psf/@PointSpreadFunction/psfROIStack_unitTest.m)
-> [smi_psf.PointSpreadFunction.scalarPSFPrasadZone_unitTest](../+smi_psf/@PointSpreadFunction/scalarPSFPrasadZone_unitTest.m)
-> [smi_psf.PointSpreadFunction.zernikeImage_unitTest](../+smi_psf/@PointSpreadFunction/zernikeImage_unitTest.m)
+  - [smi_psf.PointSpreadFunction.crlbPSFPupil_unitTest](../+smi_psf/@PointSpreadFunction/crlbPSFPupil_unitTest.m)
+  - [smi_psf.PointSpreadFunction.optimPSFZernike_unitTest](../+smi_psf/@PointSpreadFunction/optimPSFZernike_unitTest.m)
+  - [smi_psf.PointSpreadFunction.oversamplePSFPupil_unitTest](../+smi_psf/@PointSpreadFunction/oversamplePSFPupil_unitTest.m)
+  - [smi_psf.PointSpreadFunction.phaseRetrieval_unitTest](../+smi_psf/@PointSpreadFunction/phaseRetrieval_unitTest.m)
+  - [smi_psf.PointSpreadFunction.psfROIStack_unitTest](../+smi_psf/@PointSpreadFunction/psfROIStack_unitTest.m)
+  - [smi_psf.PointSpreadFunction.scalarPSFPrasadZone_unitTest](../+smi_psf/@PointSpreadFunction/scalarPSFPrasadZone_unitTest.m)
+  - [smi_psf.PointSpreadFunction.zernikeImage_unitTest](../+smi_psf/@PointSpreadFunction/zernikeImage_unitTest.m)
 - [smi_psf.Zernike.unitTest](../+smi_psf/@Zernike/unitTest.m)
 - [smi_sim.SimSMLM.unitTest](../+smi_sim/@SimSMLM/unitTest.m)
 - [smi_vis.GenerateImages.blobColorOverlay_unitTest](../+smi_vis/@GenerateImages/blobColorOverlay_unitTest.m)
