### 2023-04-18 f7adc84
Some cleanup and added Eric as an author.
