### 2023-04-25 0547eaa
Added various tests for many classes.
