+++ doc/ExtExamples/SMLM.md
@@ -1,0 +1,102 @@
+### SMLM GUI
+
+A comprehensive software platform for single-molecule localization
+microscopy's raw data fitting and subsequent analysis. 
+
+#### Graphical User Interface (GUI) Parameters:
+
+The ***smite*** GUI allow advanced data analysis without any programming
+experience. In this guide, we present detailed step-by-step protocols on
+how to analyze SMLM data using ***smite***. The ***smite*** GUI can be
+initialized by typing the command `smi.SMLM()` in MATLAB. Figure 1 shows
+the main SMLM Interface. Note that most GUI elements have tool tips:
+hover with the mouse over to the control to display it. The components
+of the GUI are briefly explained here:
+
+1. **Data:**  In this part of the GUI, a data file can be imported. We provide two different functionalities for data anaysis. One is *Test fit* in which a single dataset is processed for a quick idea of how well the analysis parameters work; the other is *Full fit* which does a complete analysis of all datasets.
+
+    - FileName: Name of the raw data file you wish to analyze. Files can be selected by the button `Select File(s)`
+    - FileDir: Name of directory containing the raw data file
+    - ResultsDir: Name of directory in which results will be saved
+    - Analysis ID: Optional identifier to be tagged onto the filenames of saved results 
+    - FileType: Type of raw data file, e.g., mat or h5
+    - DataVariable: Name of variable in raw data .mat file(s) containing the data
+    - DatasetList: Array specifying the dataset number(s) to be analyzed
+    - DatasetMods: This drop-down menu allows you to include/exclude specific datasets from analysis
+    - CameraType: The type of a camera used to collect raw data. This drop-down menu includes EMCCD or sCMOS cameras
+    - CameraGain: Gain of a camera used to collect raw data
+    - CameraOffset: Offset of a camera used to collect raw data
+    - CameraReadNoise: Variance of the read-noise of a camera used to collect raw data
+    - CalibrationFilePath: Path to camera calibration file to be used; see [CalibrationFile](../FileFormats/CalibrationFile.md)
+    - RegistrationFilePath: Path to channel registration file containing a transform applied to the data.
+    - DataROI: Region-of-Interest of camera (optional)
+    - FrameRate: Acquisition frame rate of a camera used to collect raw data
+    - PixelSize: Pixel size of a camera used to collect raw data, back-projected to the objective focal plane (um)
+    - SEAdjust: Standard error inflation applied to each localization
+
+* In the GUI, we have different buttons whose functionality is provided below:
+
+    - Import SMF: This button allows you to import an SMF structure saved in a .mat file
+    - Export SMF: This button allows you to export the current settings displayed in this GUI to an SMF structure in a .mat file
+    - Reset SMF: This button allows you to reset current settings to their default values defined in the `smi_core.SingleMoleculeFitting` class
+
+![Figure 1](SMLM1.png)
+
+2. **BoxFinding:** Box finding finds and collates subregions of local maxima from a stack of 2D images. Local maxima are used as the center of ROIs if the estimated single molecule intensity is greater than MinPhotons. The fields are:
+
+    - BoxSize: Linear box size for fitting (pixels)
+    - BoxOverlap: Overlap of boxes (pixels)
+    - MinPhotons: Minimum number of photons allowed from an emitter.
+
+![Figure 2](SMLM2.png)
+
+3. **Fitting:** Fits a 2D Gaussian blob model to the each image in a stack of 2D images. The fields are:
+
+    - PSFSigma: Known or initial PSF Sigma (pixels) which can be scalar or an array [SigmaX SigmaY]
[…]

[…]
