### 2023-04-26 d929322
Fix some references; make figures smaller?
