+++ doc/DetailedOverview.md
@@ -38,12 +38,12 @@ used by SMLM (and hence Publish and SPT) as well.
 
 Corresponding examples/extended examples are presented in:
 - [\[SMLM\]](../MATLAB/examples/Example_SMLM_Basic.m)
-  [\[SMLM_ExtExample\]](Examples/SMLM.md)
+  [\[SMLM_ExtExample\]](ExtExamples/SMLM.md)
 - [\[Publish\]](../MATLAB/examples/Example_Publish.m)
   [\[Publish_ExtExample\]](ExtExamples/Publish.md)
 - [\[SPT\]](../MATLAB/examples/Example_SPT.m)
   [\[SPT_ExtExample\]](ExtExamples/SPT.md)
-- BaGoL
+- BaGoL (see below)
   [\[BaGoL_ExtExample\]](ExtExamples/BaGoL.md)
 ```
   B=BaGoL()       % create object
