+++ doc/ExtExamples/SMLM.md
@@ -26,84 +26,143 @@ of the GUI are briefly explained here:
     - Analysis ID: Optional identifier to be tagged onto the filenames
       of saved results
     - FileType: Type of raw data file, e.g., mat or h5
-    - DataVariable: Name of variable in raw data .mat file(s) containing the data
+    - DataVariable: Name of variable in raw data .mat file(s) containing
+      the data
     - DatasetList: Array specifying the dataset number(s) to be analyzed
-    - DatasetMods: This drop-down menu allows you to include/exclude specific datasets from analysis
-    - CameraType: The type of a camera used to collect raw data. This drop-down menu includes EMCCD or sCMOS cameras
+    - DatasetMods: This drop-down menu allows you to include/exclude
+      specific datasets from analysis
+    - CameraType: The type of a camera used to collect raw data. This
+      drop-down menu includes EMCCD or sCMOS cameras
     - CameraGain: Gain of a camera used to collect raw data
     - CameraOffset: Offset of a camera used to collect raw data
-    - CameraReadNoise: Variance of the read-noise of a camera used to collect raw data
-    - CalibrationFilePath: Path to camera calibration file to be used; see [CalibrationFile](../FileFormats/CalibrationFile.md)
-    - RegistrationFilePath: Path to channel registration file containing a transform applied to the data.
+    - CameraReadNoise: Variance of the read-noise of a camera used to
+      collect raw data
+    - CalibrationFilePath: Path to camera calibration file to be used;
+      see [CalibrationFile](../FileFormats/CalibrationFile.md)
+    - RegistrationFilePath: Path to channel registration file containing
+      a transform applied to the data.
     - DataROI: Region-of-Interest of camera (optional)
     - FrameRate: Acquisition frame rate of a camera used to collect raw data
-    - PixelSize: Pixel size of a camera used to collect raw data, back-projected to the objective focal plane (um)
+    - PixelSize: Pixel size of a camera used to collect raw data,
+      back-projected to the objective focal plane (um)
     - SEAdjust: Standard error inflation applied to each localization
 
 * In the GUI, we have different buttons whose functionality is provided below:
 
-    - Import SMF: This button allows you to import an SMF structure saved in a .mat file
-    - Export SMF: This button allows you to export the current settings displayed in this GUI to an SMF structure in a .mat file
-    - Reset SMF: This button allows you to reset current settings to their default values defined in the `smi_core.SingleMoleculeFitting` class
+    - Import SMF: This button allows you to import an SMF structure
+      saved in a .mat file
+    - Export SMF: This button allows you to export the current settings
+      displayed in this GUI to an SMF structure in a .mat file
+    - Reset SMF: This button allows you to reset current settings to their
+      default values defined in the `smi_core.SingleMoleculeFitting` class
 
-![Figure 1](SMLM1.png | width=50% height=50%)
-<A HREF="SMLM1.png"><IMG SRC="SMLM1.png" WIDTH=50% HEIGHT=50%>Figure 1 (HTML)</A>
+<IMG SRC="SMLM1.png" WIDTH=50% HEIGHT=50%><BR>Figure 1
 
-2. **BoxFinding:** Box finding finds and collates subregions of local maxima from a stack of 2D images. Local maxima are used as the center of ROIs if the estimated single molecule intensity is greater than MinPhotons. The fields are:
+2. **BoxFinding:** Box finding finds and collates subregions of local
+   maxima from a stack of 2D images. Local maxima are used as the center
+   of ROIs if the estimated single molecule intensity is greater than
+   MinPhotons. The fields are:
 
     - BoxSize: Linear box size for fitting (pixels)
     - BoxOverlap: Overlap of boxes (pixels)
     - MinPhotons: Minimum number of photons allowed from an emitter.
 
[…]
