### 2023-04-26 402f408
FIx some bad figure references.
