### 2023-04-26 0355096
Added figure tiles and a stub for SPT.
