+++ doc/ExtExamples/SPT.md
@@ -1,0 +1,3 @@
+### SPT (Single Particle Tracking)
+
+<IMG SRC="SPT1.png" WIDTH=50% HEIGHT=50%><BR>Figure 1. **Tracking** tab.
