### 2023-04-27 5740a52
Fixed a bug in hierBaGoL_run not returning BGL.
