+++ MATLAB/+smi/@BaGoL/BaGoL.m
@@ -373,7 +373,7 @@ classdef BaGoL < handle
        [Coords,Ind] = SEfilter(ROIs)
        Im=genSRMAPNOverlay(SMD,MAPN,XSize,YSize,PixelSize,SaveDir,XStart,YStart,RadiusScale,ScaleBarLength)
        BGL = hierBaGoL_analysis(SMD, FileNameIn, SaveDir, BaGoLParams)
-       hierBaGoL_run(Files, DataROI, Results_BaGoL, BaGoLParams, ROIs)
+       BGL = hierBaGoL_run(Files, DataROI, Results_BaGoL, BaGoLParams, ROIs)
     end
     
 end
