### 2023-04-30 7c77ed2
Update SPT.md
