+++ doc/ExtExamples/SPT.md
@@ -1,3 +1,151 @@
 ### SPT (Single Particle Tracking)
 
+Single Particle Tracking in ***smite*** allows for robust single 
+molecule tracking by creating a TrackingResults (TR) structure, which 
+itself is an array of SMD structures, where each array element is an 
+SMD structure that corresponds to a single trajectory.
+
+TR structures are organized as follows: each
+structure element corresponds to a single trajectory, i.e., TR(n)
+contains all relevant properties of the n-th trajectory. The TR
+structure is intended to carry all necessary information from an SMD
+structure (see (smi_core.SingleMoleculeData)[SMD.md]) but organized
+in a more user-friendly manner for SPT data.
+
+
+The TrackingResults structure can be implemented for subsequent downstream 
+data analysis of each trajectory contained in the TR structure, 
+e.g. estimating diffusion, HMM analysis, etc.. (see (../examples/README.md)))
+
+
+Single particle tracking in ***smite*** utilizes the SMF 
+(Single Molecule Fitting)  structure as described here: [SMLM](../ExtExamples/SMLM.md).
+
+Parameters can be modified either in a script for batch tracking of multiple
+files (see (../examples/README.md)))  or through the SPT GUI.
+
+
+#### Graphical User Interface (GUI) Parameters:
+
+The ***smite*** SPT GUI allows for advanced single particle trajectory 
+analysis  without any programming experience. In this guide, we present 
+detailed step-by-step protocols on how to crate a TR structure by either 
+creating a new SMF structure from raw data or by importing a previously 
+created SMF structure as described here: [SMLM](../ExtExamples/SMLM.md).
+
+The ***smite*** SPT GUI can be initialized by typing the command `smi.SPT()` 
+in MATLAB. 
+
+Figure 1 shows the main SPT GUI Tracking tab interface to create a TR structure. 
+
+*All other tabs for creating a SMF structure in the SPT GUI are exactly the same as in the SMLM GUI*
+
+The components for tracking single molecule data from the SMF structure 
+either from a script or through the GUI briefly explained here:
+
+(Note that most GUI elements have tool tips: hover with the mouse over to the 
+control to display it.)
+
 <IMG SRC="SPT1.png" WIDTH=50% HEIGHT=50%><BR>Figure 1. **Tracking** tab.
+
+**Tracking:**     {SPT}
+
+  Method:         Method used for tracking. CostMatrix can only be selected as of 
+                  now. (Default='CostMatrix')
+
+  D:              Diffusion Constant (Pixels^2/Frame)  - The known or anticipated 
+                  diffusion constant of emitters in the raw data. (Default=0.01).
+
+  TrajwiseD:      Use traj.-wise value for D (logical) - This allows for trajectory-wise 
+                  estimated diffusion constants when iteratively tracking. "NIterMax" or 
+                  "NIterMaxBatch" must be 2 or higher. (Default=true)
+
+  K_on:           Off to On Rate (Frame^-1) - The known or anticipated rate at which 
+                  dark emitters become  fluorescent, or return from a dark state 
+                  (e.g. rate at which out of focus emitter comes into view.) 
+                  (Default=.9)
+
+  K_off:          On to Off Rate (Frame^-1) - The known or anticipated rate at which 
+                  emitters transition to a dark state.(Default=.1)
+
+  MaxDistFF:      Maximum distance gap for frame-to-frame connection (Pixels) - The 
+                  maximum separation between localizations such that they can still 
+                  be considered candidates for the fame-to-frame connection 
+                  procedure. (Default=5)
+
+  MaxDistGC:      Maximum distance gap for Gap Closing (Pixels) - The maximum separation 
+                  between localizations such that they can still be considered candidates 
[…]
