### 2023-05-01 0701bf1
Fix Publish to not crash on older formatted .h5 files that previously worked.
