+++ doc/ExtExamples/SMLM.md
@@ -19,41 +19,41 @@ of the GUI are briefly explained here:
    analysis parameters work; the other is *Full fit* which does a complete
    analysis of all datasets.
 
-    - FileName: Name of the raw data file you wish to analyze. Files
+    - **FileName**: Name of the raw data file you wish to analyze. Files
       can be selected by the button `Select File(s)`
-    - FileDir: Name of directory containing the raw data file
-    - ResultsDir: Name of directory in which results will be saved
-    - Analysis ID: Optional identifier to be tagged onto the filenames
+    - **FileDir**: Name of directory containing the raw data file
+    - **ResultsDir**: Name of directory in which results will be saved
+    - **Analysis ID**: Optional identifier to be tagged onto the filenames
       of saved results
-    - FileType: Type of raw data file, e.g., mat or h5
-    - DataVariable: Name of variable in raw data .mat file(s) containing
+    - **FileType**: Type of raw data file, e.g., mat or h5
+    - **DataVariable**: Name of variable in raw data .mat file(s) containing
       the data
-    - DatasetList: Array specifying the dataset number(s) to be analyzed
-    - DatasetMods: This drop-down menu allows you to include/exclude
+    - **DatasetList**: Array specifying the dataset number(s) to be analyzed
+    - **DatasetMods**: This drop-down menu allows you to include/exclude
       specific datasets from analysis
-    - CameraType: The type of a camera used to collect raw data. This
+    - **CameraType**: The type of a camera used to collect raw data. This
       drop-down menu includes EMCCD or sCMOS cameras
-    - CameraGain: Gain of a camera used to collect raw data
-    - CameraOffset: Offset of a camera used to collect raw data
-    - CameraReadNoise: Variance of the read-noise of a camera used to
+    - **CameraGain**: Gain of a camera used to collect raw data
+    - **CameraOffset**: Offset of a camera used to collect raw data
+    - **CameraReadNoise**: Variance of the read-noise of a camera used to
       collect raw data
-    - CalibrationFilePath: Path to camera calibration file to be used;
+    - **CalibrationFilePath**: Path to camera calibration file to be used;
       see [CalibrationFile](../FileFormats/CalibrationFile.md)
-    - RegistrationFilePath: Path to channel registration file containing
+    - **RegistrationFilePath**: Path to channel registration file containing
       a transform applied to the data.
-    - DataROI: Region-of-Interest of camera (optional)
-    - FrameRate: Acquisition frame rate of a camera used to collect raw data
-    - PixelSize: Pixel size of a camera used to collect raw data,
+    - **DataROI**: Region-of-Interest of camera (optional)
+    - **FrameRate**: Acquisition frame rate of a camera used to collect raw data
+    - **PixelSize**: Pixel size of a camera used to collect raw data,
       back-projected to the objective focal plane (um)
-    - SEAdjust: Standard error inflation applied to each localization
+    - **SEAdjust**: Standard error inflation applied to each localization
 
 * In the GUI, we have different buttons whose functionality is provided below:
 
-    - Import SMF: This button allows you to import an SMF structure
+    - **Import SMF**: This button allows you to import an SMF structure
       saved in a .mat file
-    - Export SMF: This button allows you to export the current settings
+    - **Export SMF**: This button allows you to export the current settings
       displayed in this GUI to an SMF structure in a .mat file
-    - Reset SMF: This button allows you to reset current settings to their
+    - **Reset SMF**: This button allows you to reset current settings to their
[…]
