### 2023-05-04 723fd0a
Updated examples/unitTests to save images consistently.
