+++ MATLAB/+smi_stat/@DiffusionEstimator/unitTest.m
@@ -3,6 +3,13 @@ function success = unitTest()
 
 success = 0;
 
+SaveDir = fullfile(tempdir, 'smite', 'unitTest', 'DiffusionEstimator');
+if ~isfolder(SaveDir)
+   mkdir(fullfile(tempdir, 'smite'));
+   mkdir(fullfile(tempdir, 'smite', 'unitTest'));
+   mkdir(fullfile(tempdir, 'smite', 'unitTest', 'DiffusionEstimator'));
+end
+
 % This script demonstrates the usage of smi_stat.DiffusionEstimator.
 
 %% Single diffusing population, estimating D as the slope of the MSD.
@@ -40,6 +47,7 @@ DE.estimateDiffusionConstant();
 % Plot the MSD and fit results.
 DE.plotEnsembleMSD(axes(figure()), DE.MSDEnsemble, DE.DiffusionStruct, ...
     DE.DiffusionModel, DE.UnitFlag);
+saveas(gcf, fullfile(SaveDir, 'DE1.png'));
 
 %% Two diffusing populations, estimating D by fitting the CDF of jumps.
 % Simulate some trajectories.
@@ -74,6 +82,7 @@ DE.estimateDiffusionConstant();
 % Plot the CDF and fit results.
 DE.plotEnsembleCDFOfJumps(axes(figure()), ...
     DE.MSDEnsemble, DE.DiffusionStruct, DE.UnitFlag);
+saveas(gcf, fullfile(SaveDir, 'DE2.png'));
 
 %% Two diffusing populations, estimating D with an MLE for the jumps.
 % (If estimating standard errors, this is faster than fitting the CDF, and
