### 2023-05-04 a9084e7
Update SPT.md which Eric provided.
