+++ doc/ExtExamples/SPT.md
@@ -7,148 +7,155 @@ SMD structure that corresponds to a single trajectory.
 
 TR structures are organized as follows: each
 structure element corresponds to a single trajectory, i.e., TR(n)
-contains all relevant properties of the n-th trajectory. The TR
+contains all relevant properties of the nth trajectory. The TR
 structure is intended to carry all necessary information from an SMD
-structure (see (smi_core.SingleMoleculeData)[SMD.md]) but organized
-in a more user-friendly manner for SPT data.
+structure (see [smi_core.SingleMoleculeData](../DataStructures/SMD.md)),
+but organized in a more user-friendly manner for SPT data.
 
-
-The TrackingResults structure can be implemented for subsequent downstream 
+The TrackingResults structure can be used for subsequent downstream 
 data analysis of each trajectory contained in the TR structure, 
-e.g. estimating diffusion, HMM analysis, etc.. (see (../MATLAB/examples/README.md))
+e.g., estimating diffusion, HMM analysis, etc.
+(see [code examples](../../MATLAB/examples/README.md)).
 
+Single particle tracking in ***smite*** also utilizes the SMF 
+(Single Molecule Fitting) structure as described here:
+[SMLM](SMLM.md).
 
-Single particle tracking in ***smite*** utilizes the SMF 
-(Single Molecule Fitting)  structure as described here: [SMLM](../ExtExamples/SMLM.md).
-
-Parameters can be modified either in a script for batch tracking of multiple
-files (see (../MATLAB/examples/README.md))  or through the SPT GUI.
-
+Parameters can be modified either in a script for batch tracking of
+multiple files (see, for example,
+[Example_SPTBatch.m](../../MATLAB/examples/Example_SPTBatch.m))
+or through the SPT GUI.
 
 #### Graphical User Interface (GUI) Parameters:
 
-The ***smite*** SPT GUI allows for advanced single particle trajectory 
-analysis  without any programming experience. In this guide, we present 
-detailed step-by-step protocols on how to crate a TR structure by either 
-creating a new SMF structure from raw data or by importing a previously 
-created SMF structure as described here: [SMLM](../ExtExamples/SMLM.md).
+The ***smite*** SPT GUI allows for advanced single particle trajectory
+analysis without any programming experience. In this extended example,
+we present detailed step-by-step protocols on how to create a TR
+structure by either creating a new SMF structure from raw data or
+by importing a previously created SMF structure as described here:
+[SMLM](SMLM.md).
 
 The ***smite*** SPT GUI can be initialized by typing the command `smi.SPT()` 
 in MATLAB. 
 
-Figure 1 shows the main SPT GUI Tracking tab interface to create a TR structure. 
-
-*All other tabs for creating a SMF structure in the SPT GUI are exactly the same as in the SMLM GUI*
+Figure 1 shows the main SPT GUI Tracking tab interface to create a
+TR structure.
+*(All other tabs for creating an SMF structure in the SPT GUI are exactly the
+same as in the description of the [SMLM GUI](SMLM.md).)*
 
 The components for tracking single molecule data from the SMF structure 
-either from a script or through the GUI briefly explained here:
-
-*(Note that most GUI elements have tool tips: hover with the mouse over to the control to display it.)*
+either from a script or through the GUI is briefly explained here:
+*(Note that most GUI elements have a tool tip: hover with the mouse over
+to the control to display it.)*
 
 <IMG SRC="SPT1.png" WIDTH=50% HEIGHT=50%><BR>Figure 1. **Tracking** tab.
 
 **Tracking:**     {SPT}
 
-  **Method**:     Method used for tracking. CostMatrix can only be selected as of 
-                  now. (Default='CostMatrix')
[…]
