+++ MATLAB/+smi_psf/@PointSpreadFunction/unitTest.m
@@ -11,12 +11,12 @@ function unitTest()
       [Report]=smi_psf.PointSpreadFunction.optimPSFZernike_unitTest();
    end
 
-   fprintf('oversamplePSFPupil_unitTest ...\n');
+   fprintf('oversamplePSFPupil_unitTest (DIPimage needed) ...\n');
    try
       [Report]=smi_psf.PointSpreadFunction.oversamplePSFPupil_unitTest();
    end
 
-   fprintf('phaseRetrieval_unitTest ...\n');
+   fprintf('phaseRetrieval_unitTest  (DIPimage needed)...\n');
    try
       [PSFStruct]=smi_psf.PointSpreadFunction.phaseRetrieval_unitTest();
    end
