+++ MATLAB/run_tests.m
@@ -3,95 +3,145 @@
 % +smi
 
 fprintf('smi.SMLM.unitTest');
-smi.SMLM.unitTest()
+try
+   smi.SMLM.unitTest()
+end
 
 fprintf('smi.SPT.unitTestFFGC (frame-to-frame and gap closing processes)');
-smi.SPT.unitTestFFGC()
+try
+   smi.SPT.unitTestFFGC()
+end
 
 % +smi_cluster
 
 fprintf('smi_cluster.Clustering.unitTest');
-smi_cluster.Clustering.unitTest()
+try
+   smi_cluster.Clustering.unitTest()
+end
 
 fprintf('smi_cluster.PairCorrelation.unitTest');
-smi_cluster.PairCorrelation.unitTest()
+try
+   smi_cluster.PairCorrelation.unitTest()
+end
 
 fprintf('smi_cluster.StatisticsClustering.unitTest');
-smi_cluster.StatisticsClustering.unitTest()
+try
+   smi_cluster.StatisticsClustering.unitTest()
+end
 
 % +smi_core
 
 fprintf('smi_core.ChannelRegistration.unitTest');
-smi_core.ChannelRegistration.unitTest()
+try
+   smi_core.ChannelRegistration.unitTest()
+end
 
 fprintf('smi_core.DataToPhotons.unitTest (convert raw data in ADUs to photons');
-smi_core.DataToPhotons.unitTest()
+try
+   smi_core.DataToPhotons.unitTest()
+end
 
 fprintf('smi_core.DriftCorrection.unitTest');
-smi_core.DriftCorrection.unitTest()
+try
+   smi_core.DriftCorrection.unitTest()
+end
 
-fprintf('smi_core.FRC.unitTest (Fourier Ring Correlation)');
-smi_core.FRC.unitTest()
+fprintf('smi_core.FRC.unitTest (Fourier Ring Correlation) [DIPimage needed]');
+try
+   smi_core.FRC.unitTest()
+end
 
 fprintf('smi_core.FrameConnection.unitTest');
-smi_core.FrameConnection.unitTest()
+try
+   smi_core.FrameConnection.unitTest()
+end
 
 fprintf('smi_core.LocalizeData.unitTest (find localizations in an image stack');
-smi_core.LocalizeData.unitTest()
+try
+   smi_core.LocalizeData.unitTest()
+end
 
 fprintf('smi_core.Threshold.unitTest');
-smi_core.Threshold.unitTest()
+try
+   smi_core.Threshold.unitTest()
+end
 
 % +smi_psf
 
 % smi_psf.PointSpreadFunction.crlbPSFPupil_unitTest()
 % smi_psf.PointSpreadFunction.optimPSFZernike_unitTest()
-% smi_psf.PointSpreadFunction.oversamplePSFPupil_unitTest()
-% smi_psf.PointSpreadFunction.phaseRetrieval_unitTest()
+% smi_psf.PointSpreadFunction.oversamplePSFPupil_unitTest() [DIPimage needed]
+% smi_psf.PointSpreadFunction.phaseRetrieval_unitTest() [DIPimage needed]
 % smi_psf.PointSpreadFunction.psfROIStack_unitTest()
 % smi_psf.PointSpreadFunction.scalarPSFPrasadZone_unitTest()
 % smi_psf.PointSpreadFunction.zernikeImage_unitTest()
 fprintf('smi_psf.PointSpreadFunction.unitTest (several individual tests)');
-smi_psf.PointSpreadFunction.unitTest()
+try
+   smi_psf.PointSpreadFunction.unitTest()
+end
 
[…]
