### 2023-05-12 c4635c9
Simplified unitTests; remove those using DIPimage; add ExpectedResults
