+++ MATLAB/run_tests.m
@@ -2,66 +2,69 @@
 
 % +smi
 
-fprintf('smi.SMLM.unitTest');
+fprintf('smi.SMLM.unitTest\n');
 try
    smi.SMLM.unitTest()
 end
 
-fprintf('smi.SPT.unitTestFFGC (frame-to-frame and gap closing processes)');
+fprintf('smi.SPT.unitTestFFGC (frame-to-frame and gap closing processes)\n');
 try
    smi.SPT.unitTestFFGC()
 end
 
 % +smi_cluster
 
-fprintf('smi_cluster.Clustering.unitTest');
+fprintf('smi_cluster.Clustering.unitTest\n');
 try
    smi_cluster.Clustering.unitTest()
 end
 
-fprintf('smi_cluster.PairCorrelation.unitTest');
+fprintf('smi_cluster.PairCorrelation.unitTest\n');
 try
    smi_cluster.PairCorrelation.unitTest()
 end
 
-fprintf('smi_cluster.StatisticsClustering.unitTest');
+fprintf('smi_cluster.StatisticsClustering.unitTest\n');
 try
    smi_cluster.StatisticsClustering.unitTest()
 end
 
 % +smi_core
 
-fprintf('smi_core.ChannelRegistration.unitTest');
+fprintf('smi_core.ChannelRegistration.unitTest\n');
 try
    smi_core.ChannelRegistration.unitTest()
 end
 
-fprintf('smi_core.DataToPhotons.unitTest (convert raw data in ADUs to photons');
+fprintf( ...
+   'smi_core.DataToPhotons.unitTest (convert raw data in ADUs to photons)\n');
 try
    smi_core.DataToPhotons.unitTest()
 end
 
-fprintf('smi_core.DriftCorrection.unitTest');
+fprintf('smi_core.DriftCorrection.unitTest\n');
 try
    smi_core.DriftCorrection.unitTest()
 end
 
-fprintf('smi_core.FRC.unitTest (Fourier Ring Correlation) [DIPimage needed]');
-try
-   smi_core.FRC.unitTest()
-end
+%fprintf( ...
+%   'smi_core.FRC.unitTest (Fourier Ring Correlation) [DIPimage needed]\n');
+%try
+%   smi_core.FRC.unitTest()
+%end
 
-fprintf('smi_core.FrameConnection.unitTest');
+fprintf('smi_core.FrameConnection.unitTest\n');
 try
    smi_core.FrameConnection.unitTest()
 end
 
-fprintf('smi_core.LocalizeData.unitTest (find localizations in an image stack');
+fprintf( ...
+   'smi_core.LocalizeData.unitTest (find localizations in an image stack\n');
 try
    smi_core.LocalizeData.unitTest()
 end
 
-fprintf('smi_core.Threshold.unitTest');
+fprintf('smi_core.Threshold.unitTest\n');
 try
    smi_core.Threshold.unitTest()
 end
@@ -70,78 +73,76 @@ end
 
 % smi_psf.PointSpreadFunction.crlbPSFPupil_unitTest()
 % smi_psf.PointSpreadFunction.optimPSFZernike_unitTest()
-% smi_psf.PointSpreadFunction.oversamplePSFPupil_unitTest() [DIPimage needed]
-% smi_psf.PointSpreadFunction.phaseRetrieval_unitTest() [DIPimage needed]
 % smi_psf.PointSpreadFunction.psfROIStack_unitTest()
 % smi_psf.PointSpreadFunction.scalarPSFPrasadZone_unitTest()
 % smi_psf.PointSpreadFunction.zernikeImage_unitTest()
[…]
