### 2023-05-15 f8bddf1
More explanation on gain/offset/readnoise; small fixes.
