+++ MATLAB/+smi_core/@SingleMoleculeFitting/SingleMoleculeFitting.m
@@ -396,7 +396,7 @@ classdef SingleMoleculeFitting < matlab.mixin.Copyable
             obj.SMFFieldNotes.Data.CalibrationFilePath.Tip = ...
                 'Path to the camera calibration file to be used.';
             obj.SMFFieldNotes.Data.RegistrationFilePath.Tip = ...
-                sprintf(['Path to a channel registration file', ...
+                sprintf(['Path to a channel registration file ', ...
                 'containing a\n transform that will be applied to the data']);
             obj.SMFFieldNotes.Data.DataROI.Tip = ...
                 sprintf(['ROI of data to be analyzed (currently only ', ...
