### 2023-05-15 6dea5e0
More additions to make gain/offset really clear!
