+++ MATLAB/+smi_core/@SingleMoleculeFitting/SingleMoleculeFitting.m
@@ -48,6 +48,10 @@ classdef SingleMoleculeFitting < matlab.mixin.Copyable
     %                DatasetMods{2} = [2, 4].
     %                (cell array of int32 arrays)(Default={[]; []})
     %   CameraType:     'EMCCD','SCMOS' (Default='EMCCD')
+    %                   NOTE: The next 3 quantities (CameraGain, CameraOffset,
+    %                   CameraReadNoise) should be scalars if the CameraType is
+    %                   'EMCCD' while if 'SCMOS', square arrays taken from the
+    %                   CalibrationFile located at the CalibrationFilePath.
     %   CameraGain:     Camera Gain, scalar or image (Default=1)
     %   CameraOffset:   Camera Offset, scalar or image (Default=0)
     %   CameraNoise:    Camera readnoise, scalar or image (Default=0)
@@ -385,7 +389,11 @@ classdef SingleMoleculeFitting < matlab.mixin.Copyable
                 'strictly excluded from the analysis. This set takes\n',...
                 'precedence over the inclusion set']);
             obj.SMFFieldNotes.Data.CameraType.Tip = ...
-                'Type of camera used to collect the raw data';
+                sprintf(['Type of camera used to collect the raw data.\n', ...
+                'CameraGain, CameraOffset, CameraReadNoise) should be\n', ...
+                'scalars if the CameraType is EMCCD while if SCMOS,\n', ...
+                'square arrays taken from the file located at the ', ...
+                'CalibrationFilePath.']);
             obj.SMFFieldNotes.Data.CameraGain.Tip = ...
                 'Gain of the camera used to collect the raw data';
             obj.SMFFieldNotes.Data.CameraOffset.Tip = ...
