### 2023-05-19 0a500d9
Added more doc to the top-level README and friends.
