+++ MATLAB/ExpectedResults/README.md
@@ -7,3 +7,7 @@ noted in the following READMEs:
 
 These tests were performed on a Linux system running Ubuntu 20.04 using
 x86_64 processors running MATLAB R2021a.
+
+The tests are frequently stochastic in nature, so outputs from run to
+run will not necessarily be identical, even on the same system, but
+the `ExpectedResults` will provide a flavor of what to expect.
