### 2023-05-19 c26cd81
Added more clarification; fixed typo.
