+++ MATLAB/+smi_core/@DriftCorrection/README.md
@@ -18,6 +18,16 @@ EXAMPLE USAGE (see also unitTest):
 
 ---
 
+CITATION:
+Michael J. Wester, David J. Schodt, Hanieh Mazloom-Farsibaf, Mohamadreza
+Fazel, Sandeep Pallikkuth and Keith A. Lidke, "Robust, fiducial-free drift
+correction for super-resolution imaging", *Scientific Reports*, Volume 11,
+Article 23672, December 8, 2021, 1-14,
+[https://www.nature.com/articles/s41598-021-02850-7](https://www.nature.com/articles/s41598-021-02850-7)
+(DOI: 10.1038/s41598-021-02850-7).
+
+---
+
 ```
 properties:
    % Intra-dataset threshold (pixel)
