### 2023-05-31 #4979 @MJWester
Changes to SMITE:
**(1) Extend the documentation to cover all core functionality.**
We have added many MarkDown files to document the code. Each class and namespace now has a README.md file linked to the top-level repository README. The various READMEs list the functions/classes contained in the corresponding class/namespace. The function summaries are linked to the actual function source codes, the headers of which document usage in detail. In addition, in the doc directory we have added further details concerning the core functionality in the +smi namespace, including documentation on the 3 major data structures used, required file formats and links to a collection of examples and unit tests. All the unit tests and more than half of the examples use data simulations, so no user intervention is needed. References to existing public SMITE software and datasets is also provided in a few places.
**(2) I still cant see any indication of how to use this package? Shouldn't there be a website with some examples?**
For the core functionality, both simple and extended examples are provided, with links to the actual MATLAB code. The extended examples for SMLM and SPT are GUI oriented with comments on how to choose (or assume the default) for the various parameters.
**(3) I recommend adding a "Getting started" section too.**
Good idea.  We added this in the top-level README.
**(4) Add automated testing, e.g. a script running several test codes.**
In the original submission, there were various unit tests, but they were not well connected to the documentation.  Linked to Code Examples in the main README (as well as other places) are a list of the SMITE examples, whether they need a GPU and if they are self-contained or need a data file. The unit tests are also listed (they are by definition, self-contained). The unit tests are processed by a script "run_tests.m"; a directory of ExpectedResults (some larger files deleted to keep the distribution small) can be found in the MATLAB subdirectory of SMITE, which is where the examples directory is located as well. Hyperlinks connect the above.
