### 2023-06-15 #5563 @editorialbot
[REVIEW]: SMITE: Single Molecule Imaging Toolbox Extraordinaire (MATLAB)
**Submitting author:** @MJWester (<a href="http://orcid.org/0000-0002-3520-7605">Michael Wester</a>)
**Repository:** https://github.com/LidkeLab/smite
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
