### 2023-06-16 #5563 @MJWester
I am not sure why there is an error.  Is this our paper?  The most recent artifact produced under issues in the main repository seems to be correct, so I am not sure where to begin.
