### 2023-06-17 #5563 @bencardoen
@MJWester when you check the output of the error, it seems it crashed on a missing affiliation
```
failure': Author (Eric A. Burns) is missing affiliation (Theoj::Error)
```
https://github.com/openjournals/joss-papers/actions/runs/5274452956/jobs/9538928136#step:3:93
