### 2023-06-20 e000de2
Typo in the paper.md markup.
