### 2023-06-21 4bf59b0
Fix combineBaGoLROIs to deal with missing ROIs.
