+++ MATLAB/+smi_cluster/@ClusterInterface/combineBaGoLROIs.m
@@ -1,4 +1,5 @@
-function defineBaGoLROIs(pathnameR, filesR, pathnameB, filesB, MAPNfile)
+function combineBaGoLROIs(pathnameR, filesR, pathnameB, filesB, MAPNfile, ...
+                          keep_numbering)
 % ---------- Possibly, combine individually processed BaGoL ROIs into a single
 %            _ROIs.mat file using the _ROIs.mat file that was used to define
 %            the ROIs originally from the SR data
@@ -31,6 +32,8 @@ function defineBaGoLROIs(pathnameR, filesR, pathnameB, filesB, MAPNfile)
 %    filesB      BaGoL files to collect ROI coordinate info from
 %    MAPNfile    if false, assume BaGoL_Results_*_Results*.mat files,
 %                otherwise if true, assume MAPN_*.mat files
+%    keep_numbering retain the ROI numbering even if there are missing ROIs
+%                (which will be treated as empty) [default: true]
 %
 % OUTPUTS:
 %    Saves pathnameB/Analysis/*_BaGoL_ROIs.mat
@@ -38,15 +41,21 @@ function defineBaGoLROIs(pathnameR, filesR, pathnameB, filesB, MAPNfile)
 % Created by
 %    Michael J. Wester (2022)
 
+   if ~exist('keep_numbering', 'var')
+      keep_numbering = true;
+   end
+
    results_dir = pathnameB;
 
+   n_files = numel(filesB);
    n_cells = numel(filesR);
    k = 0;
    for cc = 1 : n_cells
       % Extract the cell number from the file name.
       c = str2num(regexprep(filesR{cc}, '^.*Cell_([0-9][0-9]).*$', '$1'));
+      cellno(c) = cc;
       % Copy the global parameters from the original SR _ROIs.mat file.
-      dataR = load(fullfile(pathnameR, filesR{c}));
+      dataR = load(fullfile(pathnameR, filesR{cc}));
       if isfield(dataR, 'ResultsFile');
          ResultsFile = dataR.ResultsFile;
       else
@@ -73,30 +82,62 @@ function defineBaGoLROIs(pathnameR, filesR, pathnameB, filesB, MAPNfile)
 
       % Gather the coordinates and standard errors from the individual
       %_ROI_*.mat files containing BaGoL processed data.
+      found = [];   % a list of found ROIs
       RoI = cell(n_ROIs, 1);
       for jj = 1 : n_labels
          for ii = 1 : n_ROIs
             k = k + 1;
-            fileB = filesB{k};
-            dataB = load(fullfile(pathnameB, fileB));
-            % Extract the ROI number and label number from the file name. 
-            i = str2num(regexprep(fileB, '^.*ROI_([0-9][0-9]).*$',   '$1'));
-            j = str2num(regexprep(fileB, '^.*Label_([0-9][0-9]).*$', '$1'));
-            d = str2num(regexprep(fileB, '^.*Cell_([0-9][0-9]).*$',  '$1'));
-            if d ~= c
-               error('Inconsistency in Cell #: expected %d, got %d', c, d);
+            if k <= n_files
+               fileB = filesB{k};
+               dataB = load(fullfile(pathnameB, fileB));
+               % Extract the ROI number and label number from the file name. 
+               i = str2num(regexprep(fileB, '^.*ROI_([0-9][0-9]).*$',   '$1'));
+               j = str2num(regexprep(fileB, '^.*Label_([0-9][0-9]).*$', '$1'));
+               d = str2num(regexprep(fileB, '^.*Cell_([0-9][0-9]).*$',  '$1'));
+               if d ~= c
[…]
