### 2023-06-21 403a335
Skip completely empty ROIs in singleCondition.
