### 2023-06-21 LidkeLab/smite#38 @bencardoen
JOSS Review: Check contributors
As part of the JOSS [review](https://github.com/openjournals/joss-reviews/issues/5563), I have to verify this

```
Contribution and authorship: Has the submitting author (@MJWester) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
```

I can see that this holds for @MJWester , but the paper.md lists 13 authors, yet the repo has 8 contributors.
Can you please clarify in the paper what contributions the other authors made (funding, support, testing, ...)

Thanks,

Ben
