### 2023-06-21 LidkeLab/smite#39 @bencardoen
JOSS Review: Clarify contribution
Part of [JOSS Review](https://github.com/openjournals/joss-reviews/issues/5563)
In the paper you write
```
 Parts of it have already been or are in the process of being published, e.g., frame connection [@Schodt_article:2021], drift correction [@Wester_article:2021], Bayesian grouping of localizations [@Fazel_article:2022a]. Applications are described in [@FrancoNitta_article:2021; @Mazloom-Farsibaf_article:2021; @Bailey_article:2022]. Typical raw image data can be found in [@Pallikkuth_data:2018].
```

Please clarify what functionality is unique to the JOSS submission, or if no functionality is unique, clarify that this is a platform/integration of the above tools.
You package a lot of features/tools/utilities, so for the reader it can be helpful to have a (tabular?) overview of what functionality is offered.

It's non-trivial for me to find out exactly what is novel between the cited contributions above and the ones documented in the [docs](https://github.com/LidkeLab/smite/blob/main/doc/CoreOverview.md)

Note that providing an optimized / restructured / re-engineered version of the above is also of value, only at this point it isn't quite clear where the novelty of contribution lies.
