### 2023-06-21 LidkeLab/smite#40 @bencardoen
JOSS Review : Submit Smite to EPFL SRM Hub
[Joss review](https://github.com/openjournals/joss-reviews/issues/5563)
Optional but recommended suggestion, have your software (I couldn't find smite there) in the EPFL SRM Hub: https://srm.epfl.ch/SoftwareList
