### 2023-06-21 LidkeLab/smite#41 @bencardoen
JOSS Review: Consider placing HDF5/data format in context of newer formats
[Joss review](https://github.com/openjournals/joss-reviews/issues/5563)

-- Note: this is an optional comment/suggestion, I don't expect you to take action on it --

These two papers propose SRM adaptations of formats, adopting these could be of interest for your suite.

```bibtex
@article{cheeseman_adaptive_2018,
	title = {Adaptive particle representation of fluorescence microscopy images},
	volume = {9},
	copyright = {2018 The Author(s)},
	issn = {2041-1723},
	url = {https://www.nature.com/articles/s41467-018-07390-9},
	doi = {10.1038/s41467-018-07390-9},
	abstract = {Modern microscopes create a data deluge with gigabytes of data generated each second, and terabytes per day. Storing and processing this data is a severe bottleneck, not fully alleviated by data compression. We argue that this is because images are processed as grids of pixels. To address this, we propose a content-adaptive representation of fluorescence microscopy images, the Adaptive Particle Representation (APR). The APR replaces pixels with particles positioned according to image content. The APR overcomes storage bottlenecks, as data compression does, but additionally overcomes memory and processing bottlenecks. Using noisy 3D images, we show that the APR adaptively represents the content of an image while maintaining image quality and that it enables orders of magnitude benefits across a range of image processing tasks. The APR provides a simple and efficient content-aware representation of fluosrescence microscopy images.},
	language = {en},
	number = {1},
	urldate = {2021-12-14},
	journal = {Nature Communications},
	author = {Cheeseman, Bevan L. and Günther, Ulrik and Gonciarz, Krzysztof and Susik, Mateusz and Sbalzarini, Ivo F.},
	month = dec,
	year = {2018},
	keywords = {Applied mathematics, Image processing},
	pages = {5160},
}
```
and
```bibtex
@misc{saalfeld_saalfeldlabn5_2022,
	title = {saalfeldlab/n5: n5-2.5.1},
	shorttitle = {saalfeldlab/n5},
	url = {https://zenodo.org/record/6578232},
	abstract = {Not HDF5},
	urldate = {2023-06-02},
	publisher = {Zenodo},
	author = {Saalfeld, Stephan and Pisarev, Igor and Hanslovsky, Philipp and Champion, Andrew and Rueden, Curtis and Bogovic, John and Kittisopikul, Mark and {jakirkham}},
	month = may,
	year = {2022},
	doi = {10.5281/zenodo.6578232},
}

```
