### 2023-06-21 LidkeLab/smite#42 @bencardoen
JOSS Review: Make tests automatic/online
[JOSS Review](https://github.com/openjournals/joss-reviews/issues/5563)

It is commendable to provide unit-tests, but I would strongly suggest making these integrated in e.g. Travis/Circle CI.
For OSS packages the CCI free minutes would be sufficient.

See https://circleci.com/developer/orbs/orb/mathworks/matlab for how to run matlab on CCI.
Compiling C also works on CCI if needed.
