### 2023-06-21 2cfa3c6
Added includeROIs and associated changes to hierBaGoL routines.
