### 2023-06-22 0389894
Fixed bug in combineBaGoLROIs where there are no missing files.
