+++ paper/paper.bib
@@ -133,6 +133,23 @@ H. Shen, L. J. Tauzin, R. Baiyasi, W. Wang, N. Moringo, B. Shuang, and C. F. Lan
 
 # =============================================================================
 
+# SMITE: gaussMLE
+@ARTICLE{Smith_article:2010,
+   AUTHOR = {Carlas S Smith and Nikolai Joseph and Bernd Rieger and Keith A
+             Lidke},
+   TITLE = {{Fast, single-molecule localization that achieves theoretically
+             minimum uncertainty}},
+   JOURNAL = {Nature Methods},
+   VOLUME = 7,
+   NUMBER = 5,
+   MONTH = {May},
+   YEAR = 2010,
+   URL = {https://www.nature.com/articles/nmeth.1449},
+   DOI = {10.1038/nmeth.1449}
+}
+
+# -----------------------------------------------------------------------------
+
 # SMITE: BaGoL
 @UNPUBLISHED{Fazel_unpub:2019,
    AUTHOR = {Mohamadreza Fazel and Michael J. Wester and Bernd Rieger and
