### 2023-07-04 #5563 @jgostick
Hi @ajinkyakadu.  Thanks for agreeing to review.  To get started, you should generate your checklist using ```@editorialbot generate my checklist``` (while logged in as yourself, obvs).  You'll then get a list of speific things to look for, and your job as reviewer is then straightforward.
