### 2023-07-05 93a31ba
Fixed a bug in the call to hierBaGoL_run with no ROI argument.
