+++ .github/workflows/ci.yml
@@ -32,7 +32,7 @@ jobs:
       - name: Run all tests
         uses: matlab-actions/run-tests@v1
         with:
-          source-folder: MATLAB
+          source-folder: MATLAB/tests
 
       # As an alternative to run-tests, you can use run-command to execute a MATLAB script, function, or statement.
       #- name: Run all tests
