### 2023-07-06 5f97579
Add with: gpu: true to ci.yml
