+++ .github/workflows/ci.yml
@@ -32,7 +32,6 @@ jobs:
       - name: Run all tests
         uses: matlab-actions/run-tests@v1
         with:
-          gpu: true
           source-folder: MATLAB
 
       # As an alternative to run-tests, you can use run-command to execute a MATLAB script, function, or statement.
