### 2023-07-12 3f2c2b8
Remove the tests directory in preparation to replace wth test_core.m
