### 2023-07-12 795bee3
Fixed some small problems with a few unitTests.
