### 2023-07-13 a860577
Add MATLAB Actions enabled CI via test_core.m
