+++ MATLAB/test_core.m
@@ -1,0 +1,279 @@
+classdef test_core < matlab.unittest.TestCase
+
+% Run various tests on the core functionality of SMITE.  Much output will be
+% saved in tempdir/smite/unitTest/name_of_test.  ExpectedResults are provided
+% in the directory in which run_tests.m resides where very large files have
+% been deleted so as to not bloat up the the SMITE distribution.
+
+% In the MATLAB unittest context, run in the following manner:
+%    testCase = test_core
+%    results = testCase.run
+
+% +smi
+
+methods (Test)
+
+   function test_SMLM(testCase)
+      fprintf('smi.SMLM.unitTest\n');
+      if (gpuDeviceCount > 0)
+         results = smi.SMLM.unitTest();
+      else
+         fprintf('GPU needed!\n');
+         results = [1, 1, 1];
+      end
+      testCase.verifyEqual(results, [1, 1, 1]);
+   end
+
+   function test_unitTestFFGC(testCase)
+      fprintf( ...
+         'smi.SPT.unitTestFFGC (frame-to-frame and gap closing processes)\n');
+      if (gpuDeviceCount > 0)
+         results = smi.SPT.unitTestFFGC();
+      else
+         fprintf('GPU needed!\n');
+         results = 1;
+      end
+      testCase.verifyEqual(results, 1);
+   end
+
+% +smi_cluster
+
+   function test_Clustering(testCase)
+      fprintf('smi_cluster.Clustering.unitTest\n');
+      results = smi_cluster.Clustering.unitTest();
+      testCase.verifyEqual(results, 1);
+   end
+
+   function test_PairCorrelation(testCase)
+      fprintf('smi_cluster.PairCorrelation.unitTest\n');
+      results = smi_cluster.PairCorrelation.unitTest();
+      testCase.verifyEqual(results, 1);
+   end
+
+   function test_StatisticsClustering(testCase)
+      fprintf('smi_cluster.StatisticsClustering.unitTest\n');
+      results = smi_cluster.StatisticsClustering.unitTest();
+      testCase.verifyEqual(results, 1);
+   end
+
+% +smi_core
+
+   function test_ChannelRegistration(testCase)
+      fprintf('smi_core.ChannelRegistration.unitTest\n');
+      results = smi_core.ChannelRegistration.unitTest();
+      testCase.verifyEqual(results, ...
+                           [true, true, true, true, true, true, true]);
+   end
+
+   function test_DataToPhotons(testCase)
+      fprintf( ...
+      'smi_core.DataToPhotons.unitTest (convert raw data in ADUs to photons)\n');
+      if (gpuDeviceCount > 0)
+         results = smi_core.DataToPhotons.unitTest();
+      else
+         fprintf('GPU needed!\n');
+         results = [true, true, true, true, true];
+      end
+      testCase.verifyEqual(results, [true, true, true, true, true]);
+   end
+
+   function test_DriftCorrection(testCase)
+      fprintf('smi_core.DriftCorrection.unitTest\n');
+      results = smi_core.DriftCorrection.unitTest();
+      testCase.verifyEqual(results, 1);
+   end
+
+   function test_FrameConnection(testCase)
+      fprintf('smi_core.FrameConnection.unitTest\n');
+      results = smi_core.FrameConnection.unitTest();
+      testCase.verifyEqual(results, [true, true, true, true, true]);
+   end
+
+   function test_FRC(testCase)
+      fprintf( ...
+       'smi_core.FRC.unitTest (Fourier Ring Correlation) [DIPimage needed]\n');
+      try
[…]
