### 2023-07-13 d150457
Updated paper; fixed even more GPU requiring tests.
