+++ paper/paper.bib
@@ -96,7 +96,6 @@
 # -----------------------------------------------------------------------------
 
 # Single Particle Tracking (SPT) references.
-H. Shen, L. J. Tauzin, R. Baiyasi, W. Wang, N. Moringo, B. Shuang, and C. F. Landes, “Single particle tracking: From theory to biophysical applications,” Chem. Rev. 117(11), 7331–7376 (2017).
 @ARTICLE{Shen_article:2017,
    AUTHOR = {Hao {Shen} and Lawrence J. {Tauzin} and Rashad {Baiyasi} and
              Wenxiao {Wang} and Nicholas {Moringo} and Bo {Shuang} and Christy
@@ -279,6 +278,24 @@ H. Shen, L. J. Tauzin, R. Baiyasi, W. Wang, N. Moringo, B. Shuang, and C. F. Lan
                emitter density estimates, and a kinetic model for blinking
                which is used to optimize connections within a group of
                spatiotemporally colocated localizations.}
+}
+
+# -----------------------------------------------------------------------------
+
+% Diffusion Estimation.
+@ARTICLE{Relich_article:2016,
+   AUTHOR = {Peter K. {Relich} and Mark J. {Olah} and Patrick J. {Cutler} and
+             Keith A. {Lidke}},
+   TITLE = {{Estimation of the diffusion constant from intermittent
+             trajectories with variable position uncertainties}},
+   JOURNAL = {Physical Review E},
+   VOLUME = 93,
+   PAGES = {042401},
+   MONTH = {April},
+   NUMBER = 4,
+   YEAR = 2016,
+   URL = {https://link.aps.org/doi/10.1103/PhysRevE.93.042401},
+   DOI = {10.1103/PhysRevE.93.042401}
 }
 
 # =============================================================================
