+++ paper/paper.bib
@@ -377,7 +377,7 @@
    MONTH = {February 2},
    YEAR = 2018,
    PAGES = {532a},
-   DOI = {10.1016/j.bpj.2017.11.2912"}
+   DOI = {10.1016/j.bpj.2017.11.2912}
    }
 
 # =============================================================================
