### 2023-07-13 LidkeLab/smite#38 @MJWester
The development of the SMITE software has taken place over several years,
including an early period before our use of GitHub.  Each of the authors
has made a contribution to SMITE which I summarize below (and will now
include in the paper).  In the process of making this list, we added three
more authors as we think it is important that anyone that developed software
for SMITE be recognized.  I hope that this list is not too long, but SMITE
is a big package with many pieces and many contributors.

# Author Contributions

KAL conceived and supervised development of SMITE and its predecessors.
DJS, MF, HMF, MBMM and KAL coded SR localization techniques.
SP, MJW, MF and HMF implemented thresholding.
DJS and HMF wrote frame connection.
MJW, KAL and DJS developed drift correction based on an earlier version by FF.
MF and KAL devloped BaGoL; MJW wrote the interface to SMITE.
DJS, HMF, WKK, DAR and EJ designed and wrote code for single particle tracking
based on ideas from PKR.
DJS, MBMM, HMF and SP created visualizations.
SK, MJW and DJS added SMLM and SPT simulations.
KAL, SL and MJW developed Zernike polynomial point spread function engineering.
MJW packaged clustering techniques, and
SLS and MJW wrote code for various cluster statistics.
DJS added channel registration, various statistics, dimer hidden Markov
modeling and the batch Publish class.
PKR and MJO developed the diffusion estimator, while MJO also contributed to
the change detector.
MJW, SK and EAB added GitHub documentation.
MJW and DJS wrote the manuscript.
All authors reviewed the manuscript.
