### 2023-07-13 LidkeLab/smite#39 @MJWester
Thank you for the comment.  We were trying to imply that most of the SMITE
implementation is original to the JOSS submission, although some of the
algorithms, but generally not the code, have been previously published.  In
particular, single particle tracking, change detection analysis, dimer hidden
Markov modeling and the overall integration into a software package using
standard data structures are unique to the JOSS submission.  The quoted text
has been rewritten as follows:

Some of the algorithms have already been published: 2D Gaussian blob maximum
likelihood estimate [@Smith_article:2010], frame connection
[@Schodt_article:2021], drift correction [@Wester_article:2021], Bayesian
grouping of localizations [@Fazel_article:2022a], diffusion estimation
[@Relich_article:2016].  However, this is the first time that they have been
integrated together, sharing common data structures.
Applications are described in [@FrancoNitta_article:2021;
@Mazloom-Farsibaf_article:2021; @Bailey_article:2022].
Typical raw image data can be found in [@Pallikkuth_data:2018].
A summary of the namespaces and classes in SMITE can be found in the online
documentation at
[https://github.com/LidkeLab/smite/blob/main/doc/SMITEclasses.md](https://github.com/LidkeLab/smite/blob/main/doc/SMITEclasses.md).
