### 2023-07-13 LidkeLab/smite#40 @MJWester
Thank you for the suggestion.  We contacted

   daniel.sage@epfl.ch

who seems to be in charge of and contact for this website with subject:
"Pre-registration to Single-Molecule Localization Microscopy Challenge"
as directed by the EPFL website.  We are awaiting a response, although we
note that the above-mentioned list has not been updated since 2019, and
registration for the 3D software challenge (SMITE is a 2D code),
"is currently on hold and may resume shortly."

If I haven't heard from Daniel Sage when I am back from traveling (I'll be
gone until early August), I'll try contacting him again.
