### 2023-07-13 LidkeLab/smite#41 @MJWester
Thank you.  We have been considering new formats, such as N5, as another option
for storage of our datasets.  However, before we can really add this to SMITE,
we must add it to our camera acquisition software, so it's a process that
involves some effort.  But we are definitely thinking along these lines!
