### 2023-07-13 LidkeLab/smite#42 @MJWester
Thank you for the suggestion.  We have installed CI using GitHub Actions, and
are currently integrating our unitTests into this framework.  However, many of
our tests require a GPU, which is not available on a GitHub-Hosted Runner.  Our
network configuration consists of various Windows and a small number of Linux
systems, so we are trying to best decide how to go about adding CI while
providing flexibility of system usage to our developers.  Currently, we are
skipping tests if they require a GPU and one is not available.

I will be traveling for 3-weeks (back early August), so will revisit this issue
on my return, for example, moving to a different CI, but this should be much
easier now that we got the hang of this process.
