### 2023-07-15 LidkeLab/smite#38 @bencardoen
Thank you for clarifying the contributions and listing credit in detail.
