### 2023-07-15 LidkeLab/smite#40 @bencardoen
Thank you for the clarification, perhaps it'd be an idea for the community to host a global list of SMLM software (I know of other packages not listed as well). If you agree I could start a github gist/table.
