### 2023-07-15 LidkeLab/smite#39 @bencardoen
Thank you for clarifying this, with the new version the contribution is clear (and very useful for the field).
