### 2023-07-15 #4979 @bencardoen
@jgostick From my point of view all the issues I raised are addressed. The last open issue is/was integrating tests so they run online, but that is actively in progress and not a hard requirement for acceptance (the tests run locally and it's non trivial to have gpu testing for OSS).
For me, the paper is ready for acceptance. I'll monitor this and the other thread for updates if needed, but for me the checklist and review is complete and I'd recommend acceptance, the software provides valuable functionality to the SMLM community. (congrats @MJWester & team).
