+++ .github/workflows/cat_src.yml
@@ -1,0 +1,41 @@
+name: Concatentate Source and Readme
+
+on: 
+  push:
+    branches:
+      - main
+
+jobs:
+
+  build:
+
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v2
+
+    - name: Concatenate Files
+      run: |
+        cd ${{ github.workspace }}
+        find . -name '*.m' -o -name '*.md' | while read -r file; do
+          echo "$file" >> all_source.txt
+          cat "$file" >> all_source.txt
+          echo >> all_source.txt
+        done
+
+    - name: Setup Git  
+      run: |
+        git config --global user.name 'Your Name'
+        git config --global user.email 'your-email@domain.com'
+      
+    - name: Commit Files
+      run: |
+        git checkout -b gh-pages
+        git add all_source.txt
+        git commit -m "Updating all source file"
+        
+    - name: Push Changes
+      uses: ad-m/github-push-action@master
+      with:
+        github_token: ${{ secrets.GITHUB_TOKEN }}
+        branch: gh-pages
