+++ paper/paper.md
@@ -1,181 +1,0 @@
----
-title: 'SMITE: Single Molecule Imaging Toolbox Extraordinaire (MATLAB)'
-tags:
-  - MATLAB
-  - single molecule localization microscopy (SMLM)
-  - single particle tracking (SPT)
-  - super resolution
-authors:
-  - name: David J. Schodt
-    orcid: 0000-0002-8986-2736
-    affiliation: 1
-    equal-contrib: true
-  - name: Michael J. Wester
-    orcid: 0000-0002-3520-7605
-    affiliation: "1, 2" # (Multiple affiliations must be quoted)
-    equal-contrib: true
-  - name: Mohamadreza Fazel
-    orcid: 0000-0002-6215-1336
-    affiliation: 1
-  - name: Sajjad Khan
-    orcid: 0000-0002-6910-5199
-    affiliation: 1
-  - name: Hanieh Mazloom-Farsibaf
-    orcid: 0000-0002-2571-0418
-    affiliation: 1
-  - name: Sandeep Pallikkuth
-    affiliation: 1
-  - name: Marjolein B. M. Meddens
-    orcid: 0000-0002-9965-1342
-    affiliation: 1
-  - name: Farzin Farzam
-    affiliation: 1
-  - name: Eric A. Burns
-    orcid: 0000-0002-1625-2400
-    affiliation: 3
-  - name: William K. Kanagy
-    orcid: 0000-0002-5756-9965
-    affiliation: 3
-  - name: Derek A. Rinaldi
-    orcid: 0009-0000-8394-3626
-    affiliation: 3
-  - name: Elton Jhamba
-    orcid: 0000-0002-5272-6466
-    affiliation: 3
-  - name: Sheng Liu
-    affiliation: 1
-  - name: Peter K. Relich
-    orcid: 0000-0002-6063-6233
-    affiliation: 1
-  - name: Mark J. Olah
-    affiliation: 1
-  - name: Stanly L. Steinberg
-    affiliation: 2
-  - name: Keith A. Lidke^[corresponding author]
-    orcid: 0000-0002-9328-4318
-    affiliation: 1
-    corresponding: true
-affiliations:
- - name: Department of Physics and Astronomy, University of New Mexico
-   index: 1
- - name: Department of Mathematics and Statistics, University of New Mexico
-   index: 2
- - name: Department of Pathology, University of New Mexico Health Sciences Center
-   index: 3
-date: 29 September 2022
-bibliography: paper.bib
-
----
-
-# Summary
-
-Fluorescence single molecule imaging comprises a variety of techniques that
-involve detecting individual fluorescent molecules.  Many of these techniques
-involve localizing individual fluorescent molecules with precisions below the
-diffraction limit, which limits the spatial resolution of (visible) light-based
-microscopes.  These methodologies are widely used to image biological
-structures at the nanometer scale by fluorescently tagging the structures of
-interest, elucidating details of the biological behavior observed.
-
-Two common techniques are single-molecule localization microscropy (SMLM),
-[@Lidke_article:2005; @Betzig_article:2006; @Rust_article:2006;
-@Hell_article:2007; @vandeLinde_article:2011; @Fazel_article:2022]
-which is used to produce 2D or 3D super-resolution images of static or nearly
-static structures, and single-particle tracking (SPT) [@Shen_article:2017],
-which follows the time course of one or a very small number of moving tagged
-molecules.  SMLM often involves distributions of particles at medium to high
-density, while SPT works in a very low density domain.  These procedures all
-require intensive numerical computation, and the methods are tightly
-interwoven.
-
-# Statement of need
-
-SMITE is a MATLAB-based toolbox that provides analysis tools for fluorescence
[…]
