### 2023-08-07 LidkeLab/smite#40 @MJWester
I wrote again to Daniel Sage and got this reply today:

Dear Michael,
Sorry for this late answer.

We are revamping the website srm.epfl.ch and moving on a new server as
requested by our security IT department. It takes more time than expected!

ASAP, we will inform when we will update our website, and for sure we will
include your software SMITE.

Best,

Daniel
