### 2023-08-08 LidkeLab/smite#42 @MJWester
A if no gpu: then mask test y would be perfectly fine.

is basically how I set up the automated tests requiring a GPU:

   if gpuDeviceCount > 0
      do test needing a GPU
   else
      skip test
   end

Installing a local runner is more complicated, so we will mull on that, especially as it requires a spare machine as you note.  And although not automated, we can run the GPU tests manually, at least for now (the first two unitTests are the most important as they exercise most of the basic functionality).  Certainly, for future projects, we will set up automated testing.
