### 2023-08-09 LidkeLab/smite#40 @bencardoen
@MJWester thank you for sharing this, I'll resend our emails on adding to EPFL's table, best to keep these things in 1 place.
