### 2023-08-10 4114ab6
Remove hard-coded paths as much as possible from the Examples.
