+++ MATLAB/examples/Example_SMLM_Basic.m
@@ -14,12 +14,8 @@ SMLMobj=smi.SMLM()  %no input pops open a gui
 %% Scripted Analysis
 
 %or set it manually:
-SMLMobj.SMF.Data.FileDir='Y:\Sandeep\20-11-2020-DNA_PAINT_Tubulin';
-SMLMobj.SMF.Data.FileName='Dock2-Cell1-2020-11-12-10-29-58.h5';
+%SMLMobj.SMF.Data.FileDir='Y:\Sandeep\20-11-2020-DNA_PAINT_Tubulin';
+%SMLMobj.SMF.Data.FileName='Dock2-Cell1-2020-11-12-10-29-58.h5';
 
 %give a Analysis ID
-SMLMobj.SMF.Data.FileName='Dock2-Cell1-2020-11-12-10-29-58.h5';
-
-
-
-
+%SMLMobj.SMF.Data.AnalysisID='ID';
