### 2023-08-14 6fe714d
Add MacOS mex files; make general README clearer wrt mex/CUDA.
