### 2023-09-06 #5563 @jgostick
@MJWester, good news.  You and the reviewers have worked through the all the issues and tightened up this package considerably.  It now seems like it's ready to accept.  There are a few administrative steps still, like minting DOI's and whatnot.  I will generate a checklist for this in a subsequent comment.  The hard part is over now, yea!
