### 2023-09-07 5053d2d
Add one citation and 3 orcids to the paper in preparation for a new release.
