+++ paper/paper.bib
@@ -348,6 +348,21 @@
    DOI = {10.3791/63665}
 }
 
+# Trust regions (integrated into SMITE).
+@ARTICLE{Schodt_article:2023,
+   AUTHOR = {David J. {Schodt} and Farzin {Farzam} and Sheng {Liu} and
+             Keith A. {Lidke}},
+   TITLE = {{Automated multi-target super-resolution microscopy with trust
+             regions}},
+   JOURNAL = {Biomedical Optics Express},
+   VOLUME = 14,
+   NUMBER = 1,
+   YEAR = 2023,
+   PAGES = {429--440},
+   URL = {https://doi.org/10.1364/BOE.477501},
+   DOI = {10.1364/BOE.477501}
+}
+
 # Typical image datasets.
 @MISC{Pallikkuth_data:2018,
    AUTHOR = {Sandeep {Pallikkuth} and Cheyenne {Martin} and Farzin {Farzam} and
