+++ MATLAB/+smi_helpers/versionSMITE.m
@@ -1,7 +1,7 @@
 function version = versionSMITE()
 % The version (or release) of SMITE currenly being used.
 
-   version = 'development';
-   %version = '0.3.0';
+   %version = 'development';
+   version = '1.0.0';
 
 end
