### 2023-09-08 #5563 @MJWester
@jgostick  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.  This is in the top-level README?  I think we made a mistake and am attempting to fix it.  The software is currently version v1.0.0 and the DOI, which has a weird citation, is 10.5281/zenodo.8327410 .  We tried to follow the instructions https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content , but the corresponding author and I did things separately producing the 1.0.0 release and allowing zenodo to produce the DOI, resulting in some confusion.  Any advice?  I see in another repository a CITATION.cff file, so I am guessing I need to make one of these???
