### 2023-09-08 #5563 @MJWester
@jgostick , OK.  I am still confused about how I set the archive authors.  This is in zenodo or GitHub or ?  (As you can see, I have never done this before.). Is there a tutorial I should look at?

Another question, I am making  a GitHub file for grant info for our work.  It's kind of long.  Can this be added to the JOSS paper or should I refer to it as an URL in the GitHub repo?
