### 2023-09-08 #5563 @MJWester
This is the text I wish to add for grant info:
This work was supported by NIH grants
P30CA118100, R01CA248166, R01GM109888, R21GM104691, R21GM132716, RR024438,
1R01GM140284, 1R21EB019589, 5P50GM085273;
NSF grants 0954836 and 2039517; DOE grant DE-SC0019267;
support from the University of New Mexico Office of the Vice President for
Research Program for Enhancing Research Capacity; and
supported by grants from NVIDIA and utilized an NVIDIA A6000 GPU.
