### 2023-09-08 #5563 @MJWester
You can ignore these questions.  (1) I am beginning to figure things out with some local help, who suggested a .zenodo.json file in the github top-level directory. (2) I'll put in an acknowledgments section.  I guess it is not that long.
