### 2023-09-09 #5563 @jgostick
A zenodo file in the repo does not sound right.  I have also only done this once myself, but I think it's as simple as going to zenodo and entering the author info..  Maybe you're over thinking it and trying to get zenodo to do it automatically?
