### 2023-09-11 #5563 @MJWester
No.  The .zenodo.json file automatically provides correct metatdata for all releases of the repository to zenodo.  Entering author info manually to zenodo is the usual way, but with a large number of co-authors in my case, it is very error prone.  This method was recommended by our electronic information specialist.  For an example, see the .zenodo.json file in my repository (DOI to follow) and for further details, see https://developers.zenodo.org/#add-metadata-to-your-github-repository-release

Anyway, the author list and license are now consistent.  The correct DOI is 10.5281/zenodo.8336392
The version is v1.0.0 of September 11 (there is a version from September 7 from my first incorrect try with a different DOI).
