### 2023-09-12 #5563 @MJWester
Alright, I decided to just add the note and make one final release.

Same version v1.0.0 as just editing meta-data.  DOI is 10.5281/zenodo.8336799

Sorry for the extra DOIs---but now this is stable unless you find a problem.
