### 2023-09-14 #5563 @MJWester
@jgostick Acknowledgements updated.  The all version DOI is 10.5281/zenodo.8327409 (for version v1.0.0).  Hopefully, no further edits (unless you request some) will be needed.
