### 2023-09-15 #5563 @bencardoen
@jgostick I think there's an issue with the ORCID for 1 author
```
 Problem with ORCID ( 0000-0002-9939-1923) for Farzin Farzam. ORCID looks to be the wrong length (Theoj::Error)
```
