### 2023-09-21 08c33f4
Fixed bug with FitType XYNBSXSY and added CC3.0 CUDA ptx files.
