+++ MATLAB/+smi_sim/@GaussBlobs/gaussBlobROIStack.m
@@ -118,6 +118,11 @@ for ii = 1:Nloops
     if ~isempty(SMD.PSFSigma)
         xSigma = single(SMD.PSFSigma(startFr:endFr,2));
         ySigma = single(SMD.PSFSigma(startFr:endFr,1));
+    elseif ~isempty(SMD.PSFSigmaX) && ~isempty(PSFSigmaY)
+        xSigma = single(SMD.PSFSigmaX(startFr:endFr,2));
+        ySigma = single(SMD.PSFSigmaY(startFr:endFr,1));
+    else
+        error('SMD.PSFSigma and (SMD.PSFSigmaX and/or Y) are empty!');
     end
 
     Covar = single(Covariance(startFr:endFr));
