+++ MATLAB/ptx/SMITE_ptx_CC3.0/README.txt
@@ -1,0 +1,1 @@
+CUDA files compiled for GPU ComputeCapability 3.0
