### 2023-09-22 53a7e9c
Fixes bug with FitType XYNBSXSY.
