### 2023-09-26 4062ad9
Remove duplicate bib entry from paper.
