+++ paper/paper.bib
@@ -379,24 +379,6 @@
 
 # =============================================================================
 
-# Instrument control
-@ARTICLE{Pallikkuth_article:2018,
-   AUTHOR = {Sandeep {Pallikkuth} and Marjolein {Meddens} and Mohamad {Fazel}
-             and Hanieh {Farsibaf} and Farzin {Farzam} and Michael {Wester} and
-             Keith {Lidke}},
-   TITLE = {{A MATLAB-based Instrument Control Package for Fluorescence
-             Imaging}},
-   JOURNAL = {Biophysical Journal},
-   VOLUME = 114,
-   NUMBER = {3, Supplement 1},
-   MONTH = {February 2},
-   YEAR = 2018,
-   PAGES = {532a},
-   DOI = {10.1016/j.bpj.2017.11.2912}
-   }
-
-# =============================================================================
-
 # Other integrated software
 @ARTICLE{Ries_article:2020,
    AUTHOR = {Jonas {Ries}},
