### 2023-09-26 #5563 @MJWester
Except for the JOSS DOI which I didn't expect to work until formally accepted, the article I downloaded looks great.  Please let me know if I am specifically supposed to do anything else (I am assuming the editoriabotl comments above are for the editors).
