### 2023-09-30 fea14ef
Updated simplePairCorr to more rationally set directories.
