+++ MATLAB/+smi_helpers/versionSMITE.m
@@ -2,6 +2,7 @@ function version = versionSMITE()
 % The version (or release) of SMITE currenly being used.
 
    %version = 'development';
-   version = '1.0.0';
+   %version = '1.0.0';
+   version = '1.0.1';
 
 end
