### 2023-10-01 #5563 @Kevin-Mattheus-Moerman
@jgostick have you tried `@editorialbot create post-review checklist` to help remember these final steps? I'll run it now and you'll see it generates a helpful list.
