### 2023-10-01 #5563 @Kevin-Mattheus-Moerman
@MJWester I am the AEiC on this track and here to help process final steps. I have checked this review, the paper, your repository, and the archive link, and most seems in other. I only have the below minor points that require your attention:

- [ ] Check `microscropy` which should be `microscopy`
- [ ] Check `devloped` which should read `developed`
