+++ MATLAB/+smi_helpers/versionSMITE.m
@@ -2,7 +2,6 @@ function version = versionSMITE()
 % The version (or release) of SMITE currenly being used.
 
    %version = 'development';
-   %version = '1.0.0';
-   version = '1.0.1';
+   version = '1.0.0';
 
 end
