### 2023-10-01 #5563 @MJWester
@Kevin-Mattheus-Moerman  I have made the fixes above and rereleased v1.0.0 with these changes.
