### 2023-10-04 #5563 @MJWester
@Kevin-Mattheus-Moerman I don't see the revised affiliations when I look at the regenerated PDF via the JOSS DOI above.  I didn't create a new GitHub version v1.0.0 this time with the changes as I have done for earlier modifications, which then creates a new Zenodo DOI.  Should I do this or is the problem somewhere else (or the way I am looking at the paper)?
