### 2023-10-04 #5563 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
